% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-utils.R
\name{absolute_path}
\alias{absolute_path}
\title{Get the absolute path of a file name}
\usage{
absolute_path(filename)
}
\arguments{
\item{filename}{\code{\link{character}} string - file name}
}
\value{
\code{\link{character}} string - absolute path of file name
}
\description{
Get the absolute path of a file name
}
\examples{
  make_filename<- function() {
      if (.Platform$OS.type=="windows") {
          filename<- "u:/foo/bar.csv"
      }
      else if (.Platform$OS.type=="unix") {
          filename<- "~/foo/bar.csv"
      }
      else {
          stop(sprintf("\%s not recognized", .Platform$OS.type))
      }

      filename
  }
  (filename<- make_filename())
  absolute_path(filename)

}
\seealso{
\code{\link{path}}, \code{\link{dirname}}, \code{\link{basename}}
}
\author{
Thomas P. Harte
}
\keyword{\code{\link{basename}}}
\keyword{\code{\link{dirname}},}
\keyword{\code{\link{path}},}
