% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/businessDates.R
\name{addBusinessTimeUnits}
\alias{addBusinessTimeUnits}
\title{Add business time units}
\usage{
addBusinessTimeUnits(dates, n, timeUnit)
}
\arguments{
\item{dates}{Vector of Date objects}

\item{n}{Number of units to add}

\item{timeUnit}{Units of time: 0 = days, 1 = weeks, 2 = months, 3 = years}
}
\description{
Given a vector of Date objects,
add or subtract days, weeks, months, or years
according to the calendar of business dates.
}
