% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.xts.tbl_df.R
\name{as.xts.tbl_df}
\alias{as.xts.tbl_df}
\title{Convert a tibble to xts}
\usage{
as.xts.tbl_df(x, index = NULL)
}
\arguments{
\item{x}{A tibble to be converted to xts}

\item{index}{(character)}
}
\description{
Convert a tibble to xts
}
