% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object-utils.R
\name{bind_to_env}
\alias{bind_to_env}
\title{Bind an object to an environment}
\usage{
bind_to_env(object, envir)
}
\arguments{
\item{object}{of any type}

\item{envir}{\code{\link{environment}} to which to bind \code{object}}
}
\value{
\code{\link{invisible}}
}
\description{
Bind an object to an \code{\link{environment}}. If the object is a
\code{\link{function}} then the environment of the
\code{\link{function}} is also set to the \code{\link{environment}}.
}
\examples{
   e<- new.env()

   x<- 1:3
   `fun`<- function() {
       x
   }

   bind_to_env(x, env=e)
   bind_to_env(fun, env=e)
   x<- 1:10

   all.equal(
       fun(), 1:10
   )

   whos(sort="Name", env=e, omit=NULL)

}
\seealso{
\code{\link{environment}}, \code{\link{function}}
}
\author{
Thomas P. Harte
}
\keyword{\code{\link{environment}},}
\keyword{\code{\link{function}}}
