% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ligne.R
\name{build_line}
\alias{build_line}
\title{build_line function}
\usage{
build_line(
  type = "",
  sym = "",
  strike = 0,
  expdate = "",
  DTE = NA,
  startPrice = 0,
  endPrice = 0,
  p_u_price = 0,
  startPriceIV = 0,
  endPriceIV = 0,
  pos = 0,
  mul = NA,
  delta = NA,
  deltanotional = NA
)
}
\arguments{
\item{type}{Put Call or Stock}

\item{sym}{ticker name Empty default value}

\item{strike}{strike value - 0 is default}

\item{expdate}{expiration date Empty default value}

\item{DTE}{number of days to expiration -NA is default}

\item{startPrice}{initial price - 0 is default}

\item{endPrice}{final price - 0 is default}

\item{p_u_price}{projected underlying price - 0 is default}

\item{startPriceIV}{implied vol for initial price - 0 is default}

\item{endPriceIV}{implied vol for final price - 0 is default}

\item{pos}{position  - 0 is default}

\item{mul}{multiplier - NA is default}

\item{delta}{delta value  - NA is default}

\item{deltanotional}{delta notional value (delta dollars)  - NA is default}
}
\value{
a data frame with the following arguments: instrument,type,strike,mul,pos,startPrice,startPriceIV,
endPrice,endPriceIV, change= (endPrice-startPrice)/startPrice, startValue=mul\emph{pos}startPrice,
endValue=mul\emph{pos}endPrice, unrealizedPnL=mul\emph{pos}(endPrice-startPrice), DTE, delta, deltanotional.
NB: If mul is NA then startValue, endValue, unPnL are all equal to NA.
}
\description{
This utility transforms a set of arguments into a dataframe. Any missing argument gets a default value,
}
\examples{
build_line()
build_line(type="Stock",startPrice=100,endPrice=110,mul=1,pos=5,delta=5,deltanotional = 5*110)
}
