% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{buildInstrumentName}
\alias{buildInstrumentName}
\title{buildInstrumentName}
\usage{
buildInstrumentName(sym, expdate = as.Date("2023-01-01"), strike = 100, type)
}
\arguments{
\item{sym}{is the underlying name (e.g. SPY)}

\item{expdate}{is the expiration date (e.g. Dec 15, 2023)}

\item{strike}{is the strike number (e.g. 400)}

\item{type}{is character vector - should be one of the following: "P", "C", "Put", "Call"}
}
\description{
Objective is to build a string corresponding to a specific option (using IBKR TWS).
The string can then be searched in any file generated by IBKR website.
This function performs some checks (for instance is expdate a date, is type a valid type)
but no extensive checks are performed.
\if{html}{\out{<br>}} If type is unknown (neither Put nor Call nor P nor C) then sym is returned.
}
\examples{
buildInstrumentName("SPY",as.Date("2023-12-15"),400.0,"Put")
}
\keyword{Black-Scholes}
\keyword{date}
\keyword{options}
