% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/businessDates.R
\name{bus-date-arith}
\alias{bus-date-arith}
\alias{addBusinessDays}
\alias{subBusinessDays}
\alias{addBusinessWeeks}
\alias{subBusinessWeeks}
\alias{addBusinessMonths}
\alias{subBusinessMonths}
\alias{addBusinessYears}
\alias{subBusinessYears}
\title{Business unit date arithmetic}
\usage{
addBusinessDays(dates, n)

subBusinessDays(dates, n)

addBusinessWeeks(dates, n)

subBusinessWeeks(dates, n)

addBusinessMonths(dates, n)

subBusinessMonths(dates, n)

addBusinessYears(dates, n)

subBusinessYears(dates, n)
}
\arguments{
\item{dates}{Vector of Date objects}

\item{n}{Number ov days, weeks, months, or years
to either add or subtract (integer)}
}
\description{
Given a vector of dates, add or subtract
 business date units from it: days, weeks, months, or years.
}
