% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/businessDates.R
\name{businessCalendar}
\alias{businessCalendar}
\title{Generate vector of business dates}
\usage{
businessCalendar(from = START_OF_TIME, to = NULL)
}
\arguments{
\item{from}{Date, or character string in standard date format;
or numeric, meaning the number of days before \code{to} date}

\item{to}{Date or character string in standard date format}
}
\value{
Vector of Date objects
}
\description{
Generate vector of business dates
}
