% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/businessDates.R
\name{businessDaySeq}
\alias{businessDaySeq}
\title{Create sequence of business dates}
\usage{
businessDaySeq(from, horizon)
}
\arguments{
\item{from}{A Date object or character string in standard date format}

\item{horizon}{Either a number of dates or an end Date}
}
\value{
Vector of Date objects
}
\description{
Create sequence of business dates
}
