% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datePartFactors.R
\name{calendarDayFactor}
\alias{calendarDayFactor}
\title{Create day-of-year factor from zoo or xts object}
\usage{
calendarDayFactor(z)
}
\arguments{
\item{z}{A zoo or xts object}
}
\value{
Factor with 366 levels
}
\description{
Given an xts/zoo object with a Date index,
return a parallel factor whose levels are the
calendar day of each Date
}
