% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string-utils.R
\name{capitalize}
\alias{capitalize}
\title{Capitalize first character words in a string}
\usage{
capitalize(x)
}
\arguments{
\item{x}{\code{\link{character}} string to capitalize}
}
\value{
\code{\link{character}} string with start of words capitalized
}
\description{
Capitalize the first character of each word in a character string
}
\examples{
capitalize("hello, world!")

}
\seealso{
\code{\link{gsub}}
}
\author{
Thomas P. Harte
}
