% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{catln}
\alias{catln}
\title{Add free newline to cat arguments}
\usage{
catln(...)
}
\arguments{
\item{...}{Arguments to cat}
}
\value{
Nothing. Called for side-effect
}
\description{
Add free newline to cat arguments
}
