% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart.R
\name{chartServer}
\alias{chartServer}
\title{Chart module server}
\usage{
chartServer(id, data)
}
\arguments{
\item{id}{this is used by caller to identify chart and have the link with server piece}

\item{data}{data frame containing \code{x}, \code{y1} and \code{y2} columns}
}
\description{
This module provides the ability to display a chart with 2 curves using the same underlying
}
