% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart.R
\name{chartUI}
\alias{chartUI}
\title{Chart module UI}
\usage{
chartUI(id)
}
\arguments{
\item{id}{this is used by caller to identify chart and have the link with server piece}
}
\description{
This module provides the ability to display a chart with 2 curves using the same underlying
}
