% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{coalesce}
\alias{coalesce}
\title{Return first non-NULL, non-NA argument}
\usage{
coalesce(...)
}
\arguments{
\item{...}{Just arguments}
}
\value{
First nont-NULL, non-NA value
}
\description{
Return first non-NULL, non-NA argument
}
