% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object-utils.R
\name{col_classes}
\alias{col_classes}
\title{Get column classes in \code{\link{data.frame}}}
\usage{
col_classes(DF, flip = FALSE)
}
\arguments{
\item{DF}{\code{\link{data.frame}} or \code{\link{matrix}}}

\item{flip}{\code{\link{logical}} flips results sideways, if \code{TRUE}}
}
\value{
named \code{\link{vector}} of \code{\link{character}} strings
}
\description{
Get column classes in \code{\link{data.frame}}
}
\examples{
 DF<- data.frame(
       names=c("one","two","three"),
       numbers=1:3,
       stringsAsFactors=FALSE
  )

 out<- col_classes(DF)

 all(colnames(out)==c("names","numbers"))
 out["names"] == "character"
 out["numbers"] == "integer"

}
\seealso{
\code{\link{class}}
}
\author{
Thomas P. Harte
}
\keyword{\code{\link{class}}}
