% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commaFmt.R
\name{commaFmt}
\alias{commaFmt}
\title{Format a number with embedded commas}
\usage{
commaFmt(x, digits = 0)
}
\arguments{
\item{x}{A numeric value}

\item{Number}{of places to right of decimal point}
}
\value{
Formatted number (character)
}
\description{
Format a number with embedded commas
}
