% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condProb.R
\name{condProb}
\alias{condProb}
\title{Calculate conditional probability}
\usage{
condProb(succ, trials = NULL, ...)
}
\arguments{
\item{succ}{Vector where \code{succ[i]} is \code{TRUE}
if trial \code{i} succeeded (logical)}

\item{trials}{Mask indicating with elements of \code{succ}
are the relevent trials; NULL means all elements are relevent
(logical)}

\item{...}{Passed to \code{prop.test} function
(e.g., \code{conf.level} or \code{alternative})}
}
\value{
Returns a tibble with three columns
  \itemize{
    \item CondProb - conditional probability
    \item Low - lower limit of confidence interval
    \item High - upper limit of confidence interval
 }
}
\description{
This returns the conditional probability: Pr(succ | trial).
}
\details{
Without the \code{trials} argument, this becomes
 a simple probability calculation.

 In any event, the output include lower and upper
 confidence bounds.
}
