% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/currency.R
\name{convert_to_usd}
\alias{convert_to_usd}
\title{convert_to_usd}
\usage{
convert_to_usd(amount, currency, EUR, CHF)
}
\arguments{
\item{amount, currency}{amount is the number to be converted, currency is a string whose value is either EUR, CHF}

\item{EUR, CHF}{EUR is the value of 1 euro in USD, CHF is the value of 1 CHF in USD}
}
\description{
This function converts the amount of currency into USD, using EUR and CHF currency pairs values
}
\details{
It merely performs a multiplication of the amount by currency pair value. This function can be vectorized
}
\examples{
convert_to_usd(100.45,"EUR",1.09,1.14)
convert_to_usd(c(10000,500),c("CHF","EUR"),1.09,1.14)
convert_to_usd(c(750.543,10),c("USD","EUR"),1.09,1.14)
}
\keyword{currency}
\keyword{trading}
