% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/currency.R
\name{convert_to_usd_date}
\alias{convert_to_usd_date}
\title{convert_to_usd_date}
\usage{
convert_to_usd_date(amount, currency, date)
}
\arguments{
\item{amount, currency}{amount is the number to be converted, currency is a string whose value is either EUR, CHF}

\item{date}{date is the requested date, it must be in Date format}
}
\description{
This function converts the amount of currency into USD, using EUR and CHF currency pairs values for a given date
}
\details{
First it loads all currency pairs that have been stored for a while, then looks up for the nearest date in the CurrencyPairs table, compared with input date.
It retrieves the EUR and CHF corresponding values.

Last it calls the convert_to_usd function.

This function can be vectorized.
}
\examples{
convert_to_usd_date(100.45,"EUR",as.Date("2023-10-15"))
convert_to_usd_date(c(10000,500),c("CHF","EUR"),as.Date("2021-01-09"))
convert_to_usd_date(c(750.543,10),c("USD","EUR"),as.Date("2023-12-03"))
}
\keyword{currency}
\keyword{trading}
