% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object-utils.R
\name{copy}
\alias{copy}
\title{Copy elements of a list}
\usage{
copy(x, index = 1:length(x))
}
\arguments{
\item{x}{\code{\link{list}}, unnamed or named}

\item{index}{\code{\link{integer}} or \code{\link{character}}
index of the elements in \code{x} to copy, either indexed
\code{\link{integer}} or named \code{\link{character}}}
}
\value{
\code{\link{list}}
}
\description{
Copies elements of a list, either indexed or named
}
\examples{
 # NULL list:
	copy(list(), integer(0))

	# unnamed list:
 x<- list(1, 2, 3)
 x

 copy(x, integer(0))
 copy(x, 1:2)

 # named list:
 x<- list(a=1, b=2, c=3)
 x
 # integer index still works:
 copy(x, 1:2)
 # named index works too:
 copy(x, c("a","b"))

 ## FIXME: index mismatch
 ## copy(x, 1:4)
 ## copy(x, c("a","b"))
 ## copy(x, c("a","b","d"))

}
\seealso{
\code{\link{list}}
}
\author{
Thomas P. Harte
}
\keyword{\code{\link{list}}}
