% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object-utils.R
\name{copy_vars}
\alias{copy_vars}
\title{Copy variables from one environment to another environment}
\usage{
copy_vars(from.env, to.env = .GlobalEnv, vars = ls(envir = from.env))
}
\arguments{
\item{from.env}{\code{\link{environment}} where variables are to be copied from}

\item{to.env}{\code{\link{environment}} where variables are to be copied to}

\item{vars}{\code{\link{character}} \code{\link{vector}} of named
variables to copy from \code{\link{from.env}}}
}
\value{
\code{\link{invisible}}
}
\description{
Copy variables from one environment to another environment
}
\examples{
e<- new.env()
local({
  a<- "a"
  b<- "this is 'b'"
  x<- pi
}, env=e)

f<- new.env()
copy_vars(from=e, to=f)
whos(sort="Name", env=f)

g<- new.env()
copy_vars(from=e, to=g, vars=(c("b","x")))
whos(sort="Name", env=g)

}
\seealso{
\code{\link{put.var}}
}
\author{
Thomas P. Harte
}
\keyword{\code{\link{environment}}}
