% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-utils.R
\name{cut_system}
\alias{cut_system}
\title{Use system \code{cut} (*NIX) only}
\usage{
cut_system(cols, filename, sep = "|")
}
\arguments{
\item{cols}{\code{\link{integer}} \code{\link{vector}} index of columns to cut}

\item{filename}{\code{\link{character}} file name}

\item{sep}{\code{\link{character}} delimiter (passed to \code{cut})}
}
\value{
\code{\link{character}} \code{\link{vector}} output from \code{cut}
}
\description{
Use system \code{cut} (*NIX) only. For large files this is much faster
than reading the file into R and cutting the columns.
}
\examples{
tab<- read.csv(text='Name,Age,Salary,ID
            Dick,38,32k,1
            Tom,21,21k,2
            Harry,56,NA,3',
            header=TRUE,
            stringsAsFactors=FALSE
   )
   for (col in which(col_classes(tab)=="character"))
       tab[, col]<- tutils::trim(tab[, col])
   tab

   filename<- paste(tempfile(), ".csv", sep="")
   write.csv(tab, file=filename, row.names=FALSE, quote=FALSE)
   cut_system(1, filename, sep=",")
   cut_system(4, filename, sep=",")
   cut_system(c(1,4), filename, sep=",")
   unlink(filename)

}
\seealso{
\code{\link{system}}, \code{\link{cut}}
}
\author{
Thomas P. Harte
}
\keyword{\code{\link{cut}}}
\keyword{\code{\link{system}},}
