% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtsConversion.R
\name{dataFrameToXts_OBSOLETE}
\alias{dataFrameToXts_OBSOLETE}
\title{Convert a data frame to an xts matrix}
\usage{
dataFrameToXts_OBSOLETE(dfrm)
}
\arguments{
\item{dfrm}{A data frame with a 'date' column}
}
\value{
An xts matrix
}
\description{
Convert a data frame to an xts matrix
}
