% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateFunctions.R
\name{ddd_mm_dd}
\alias{ddd_mm_dd}
\title{Format a date concisely}
\usage{
ddd_mm_dd(d)
}
\arguments{
\item{d}{Date object}
}
\value{
Returns the formatted date (character)
}
\description{
Format a date as "MMM, DD/MM", for example "Wed, 09/24".
}
\details{
This format is useful in contexts where space is tight, readability is valuable,
and the year is understood (or not important).
}
