% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/declare.R
\name{declare}
\alias{declare}
\title{Type-check a function parameter}
\usage{
declare(..., ..env = parent.frame())
}
\arguments{
\item{...}{Parameters of the form name="type"}
}
\value{
Nothing. Called for (possible) side-effect
}
\description{
Stops if parameter type is wrong
}
