% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{display_error_message}
\alias{display_error_message}
\title{Display error message}
\usage{
display_error_message(error_msg)
}
\arguments{
\item{error_msg}{string to be displayed. Default to empty string}
}
\description{
Within any Shiny app, this function allows to display an error message
The end user then has to click OK button
The window is modal
This function cannot be used outside a Shiny app
}
\examples{
display_error_message("You must provide a valid integer!")
display_error_message("No date available!")
}
\keyword{Shiny}
\keyword{error}
