% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{display_message}
\alias{display_message}
\title{Display message}
\usage{
display_message(msg)
}
\arguments{
\item{msg}{string to be displayed. Default to empty string}
}
\description{
Within any Shiny app, this function allows to display an information message
The end user then has to click OK button
The window is modal
This function cannot be used outside a Shiny app
}
\examples{
display_message("Pi value is 3.14")
}
\keyword{Shiny}
