% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{do.rbind}
\alias{do.rbind}
\title{Bind list elements into rows}
\usage{
do.rbind(lst)
}
\arguments{
\item{lst}{A list}
}
\value{
A matrix whose rows are taken from the list elements
}
\description{
This is handy with the pipe operator:
 lapply(aList, aFunction) %>% do.rbind
}
\seealso{
\link{do.cbind}
}
