% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_if.R
\name{do_if}
\alias{do_if}
\title{Conditionally apply expressions to a data object}
\usage{
do_if(.data, condition, .then, .else = NULL)
}
\arguments{
\item{.data}{Input data}

\item{condition}{A logical value to determine whether to use .then or .else}

\item{.then}{Formula or function to apply to input data when condition is TRUE}

\item{.else}{Formula or function to apply to input data when condition is FALSE;
if NULL and condition is FALSE, data is left unchanged}
}
\value{
Output of appropriate .then/.else call
}
\description{
Conditionally apply expressions to a data object
}
