% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object-utils.R
\name{equal_tol}
\alias{equal_tol}
\title{Check that objects are equal to within some numerical tolerance}
\usage{
equal_tol(x, y, tol = 1, pct = FALSE, each = FALSE, na.rm = FALSE)
}
\arguments{
\item{x}{\code{\link{vector}}}

\item{y}{\code{\link{vector}}}

\item{tol}{\code{\link{numeric}} tolerance}

\item{pct}{\code{\link{logical}} tolerance is a percentage of mean}

\item{each}{\code{\link{logical}} returns element-by-element comparison, if \code{TRUE}}

\item{na.rm}{\code{\link{logical}} remove NAs, if \code{TRUE}}
}
\value{
\code{\link{logical}}
}
\description{
Check that objects are equal to within some numerical tolerance
}
\examples{
   x<- c(NA, 1:10)
   set.seed(1)
   y<- x + rnorm(length(x), sd=.01)
   cbind(x, y)

   equal_tol(x, y, tol=0.1, na.rm=TRUE)
   equal_tol(x, y, tol=0.1, pct=TRUE, na.rm=TRUE)
   equal_tol(x, y, tol=0.1, each=TRUE, na.rm=TRUE)
   all(equal_tol(x, y, tol=0.1, each=TRUE, na.rm=TRUE))

}
\seealso{
\code{\link{mean}}
}
\author{
Thomas P. Harte
}
\keyword{\code{\link{mean}}}
