% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fatal.R
\name{fatal}
\alias{fatal}
\alias{fatalIf}
\alias{fatalIfNot}
\title{Report or test a fatal condition}
\usage{
fatal(..., sep = " ", caller = NULL)

fatalIf(cond, ..., caller = NULL)

fatalIfNot(cond, ..., caller = NULL)
}
\arguments{
\item{...}{Passed to \code{cat}}

\item{sep}{Separator, passed to \code{cat} (character)}

\item{caller}{Name of calling function, defaults to name in caller's stack frame (character)}

\item{cond}{Boolean expression that must be FALSE (\code{fatalIf})
or must be TRUE (\code{fatalIfNot}), otherwise a \code{stop} error is signalled}
}
\description{
Report or test a fatal condition
}
