% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/businessDates.R
\name{fillBusinessDays}
\alias{fillBusinessDays}
\title{Replace missing business days}
\usage{
fillBusinessDays(dates)
}
\arguments{
\item{dates}{A vector of Dates}
}
\value{
Same vector, with trailing NAs replaced by business days
}
\description{
Given a vector of Dates with NAs at the end,
    replace the NAs with business days
}
