% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/financialFmt.R
\name{financialFmt}
\alias{financialFmt}
\title{Format a number in "financial" format}
\usage{
financialFmt(x, digits = 2)
}
\arguments{
\item{x}{Value to be formatted}

\item{digits}{Round to this number of digitis, default 2}
}
\value{
A string
}
\description{
Format a number in "financial" format
}
