% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{findNearestNumberOrDate}
\alias{findNearestNumberOrDate}
\title{findNearestNumberOrDate}
\usage{
findNearestNumberOrDate(values_list, target)
}
\arguments{
\item{values_list}{is a list of numbers or dates}

\item{target}{is a date or a number}
}
\description{
Given a list of numbers or dates,
The aim is to find the nearest number or date within the list compared with target
The list does not need to be ordered
This function will proceed by using which.min applied to list - target
}
\examples{
findNearestNumberOrDate(c(1,4,5),3)
}
\keyword{dates}
\keyword{minimum}
\keyword{numbers}
