% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object-utils.R
\name{first}
\alias{first}
\alias{last}
\alias{and}
\alias{return}
\alias{the}
\alias{(non-\code{NA})}
\alias{objects}
\alias{respectively}
\alias{using}
\alias{\code{\link{[}}}
\title{Extract first / last object}
\usage{
first(x)

last(x)
}
\arguments{
\item{x}{\code{object} indexable by \code{\link{[}}}
}
\value{
\code{object} of same \code{mode} as \code{x}
}
\description{
Extract first / last object
}
\examples{
  # from the following phrases
  str<- c("hello there", "how are you?")
  cat(paste(str), sep="\\n")
  # extract the first words:
  sapply(strsplit(str, " "), first)
  # extract the last words:
  sapply(strsplit(str, " "), last)

  # in fact this is a very useful idiom, e.g. with stock tickers:
  tickers<- c("INDU Index", "USD Curncy")
  cat(paste(tickers), sep="\\n")
  short.tickers<- sapply(strsplit(tickers, " "), first)
  short.tickers
  category<- sapply(strsplit(tickers, " "), last)
  category

}
\seealso{
\code{\link{Extract}}, \code{\link{na.omit}}
}
\author{
Thomas P. Harte
}
\keyword{\code{\link{sort}}}
