% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fwdDiff.R
\name{fwdDiff}
\alias{fwdDiff}
\title{Forward difference of a time series}
\usage{
fwdDiff(x, lag, na.pad = TRUE)
}
\arguments{
\item{x}{zoo or xts matrix}

\item{lag}{numeric; period to difference over}

\item{na.pad}{logical}
}
\value{
A matrix (same type as input) of forward differences
}
\description{
Forward difference of a time series
}
