% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcd.R
\name{gcd}
\alias{gcd}
\title{Greatest common divisor (GCD)}
\usage{
gcd(a, b)
}
\arguments{
\item{a}{An integer}

\item{b}{Another integer}
}
\value{
The greatest common divisor of \code{a} and \code{b}
}
\description{
This is used by the Intermarket Spread objects
to find the largest, common contract size.
}
\seealso{
Algorithm is descibed on
\href{https://en.wikipedia.org/wiki/Euclidean_algorithm}{Wikipedia}
}
