% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts-utils-main.R
\name{get.diff}
\alias{get.diff}
\title{One-liner: What the function does}
\usage{
get.diff(x)
}
\arguments{
\item{arg1}{\code{\link{type of arg1}} explanation of arg1}

\item{arg2}{\code{\link{type of arg2}} explanation of arg2}
}
\value{
\code{\link{return value}}
}
\description{
Description: What the function does in more detail
}
\examples{
  \dontrun{
 ## Not run:
 ## cat(sprintf("hello, world"))
  }

}
\seealso{
\code{\link{stuff1}}, \code{\link{stuff2}}
}
\author{
Thomas P. Harte
}
\keyword{\code{\link{other}
\keyword{stuff}}}
