% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts-utils.R
\name{get_diff}
\alias{get_diff}
\title{Compute the difference of values in vector}
\usage{
get_diff(x)
}
\arguments{
\item{x}{\code{\link{numeric}} \code{\link{vector}}}
}
\value{
\code{\link{numeric}} \code{\link{vector}}
}
\description{
Compute the difference of values in vector
}
\examples{
  get_diff(c(NA))
  get_diff(c(NA, NA))
  get_diff(c(NaN, NA))
  get_diff(c(NaN, 1, 1.1))
  get_diff(c(1, +Inf, 1.1))

  dates<- seq(as.Date("2011-01-01"), by=1, len=6)
  z<- zoo(c(NA, 1, 2, 3, 0, 0), order.by=dates)

  zoo::rollapply(z, 2, get_diff, align="right")
  # 2011-01-02 2011-01-03 2011-01-04 2011-01-05 2011-01-06
  #	  NA        1.0        1.0       -3.0         0

}
\seealso{
\code{\link{numeric}}, \code{\link{diff}}
}
\author{
Thomas P. Harte
}
\keyword{\code{\link{diff}}}
\keyword{\code{\link{numeric}},}
