% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string-utils.R
\name{get_file_ext}
\alias{get_file_ext}
\title{Get file extension}
\usage{
get_file_ext(x, parse = FALSE)
}
\arguments{
\item{x}{\code{\link{character}} a file name}

\item{parse}{\code{\link{logical}} if TRUE, the filename
 is parsed and returned as a \code{\link{list}}
with two elements: basename and extension}
}
\value{
\code{\link{character}} file extension
}
\description{
Returns the file extension of a file name
}
\examples{
  # "csv" :
  get_file_ext("foo.bar.csv")
  # "csv" :
  get_file_ext("foo.bar.csv", parse=TRUE)
  # $basename
  # [1] "foo.bar"
  # $extension
  # [1] "csv"

}
\seealso{
\code{\link[limma]{removeExt}}
}
\author{
Thomas P. Harte
}
\keyword{\code{\link{sort}}}
