% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts-utils.R
\name{get_return}
\alias{get_return}
\title{Compute (simple net) return of values in vector}
\usage{
get_return(x, nan.replace = FALSE)
}
\arguments{
\item{x}{\code{\link{numeric}} \code{\link{vector}}}

\item{nan.replace}{\code{\link{logical}} if TRUE, replace \code{\link{NaN}} (not-a-number) symbols with \code{NA}}
}
\value{
\code{\link{numeric}} \code{\link{vector}}
}
\description{
Compute (simple net) return of values in vector
}
\examples{
  get_return(c(NA))
  get_return(c(NA, NA))
  get_return(c(NaN, NA))
  get_return(c(NaN, 1, 1.1), nan.replace=TRUE)
  get_return(c(1, +Inf, 1.1), nan.replace=TRUE)

}
\seealso{
\code{\link{numeric}}, \code{\link{NaN}}, \code{\link{vector}}
}
\author{
Thomas P. Harte
}
\keyword{\code{\link{NaN}},}
\keyword{\code{\link{numeric}},}
\keyword{\code{\link{vector}}}
