% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-utils.R
\name{get_ticker_from_Bloomberg}
\alias{get_ticker_from_Bloomberg}
\title{Get ticker from Bloomberg.}
\usage{
get_ticker_from_Bloomberg(tickers, field = "PX_LAST", block = 5,
  start = NULL, end = Sys.Date(), file = character(0))
}
\arguments{
\item{tickers}{\code{\link{character}} Bloomberg tickers to fetch}

\item{field}{\code{\link{character}} Bloomberg data field}

\item{block}{\code{\link{integer}}
number of calendar years of data to fetch at a time
(if both \code{start} and \code{end} are \code{NULL}
then data are fetched with this block size from as
far back as possible to the latest date possible in
blocks of size \code{block})}

\item{start}{\code{\link{character}} or \code{\link{Date}}.
starting date of complete data block: if \code{NULL}
signals to fetch data as far back as possible}

\item{end}{\code{\link{character}} or \code{\link{Date}}
end date of complete data block: if \code{NULL}
signals to fetch data to the latest date possible}

\item{file}{\code{\link{character}} filename (with extension) to save data}
}
\value{
\code{\link{zoo}}
}
\description{
Returns the selected tickers from Bloomberg
}
\examples{
  \dontrun{
 ## Not run:
 ## get_ticker_from.Bloomberg("INDU INDEX", field="PX_LAST")
 ## get_ticker_from.Bloomberg("INDU INDEX", start="2008-01-01")
 ## get_ticker_from.Bloomberg("INDU INDEX", start="2008-01-01", end="2008-10-31")
  }

}
\seealso{
\code{\link{blpGetData}}, \code{\link{blpConnect}}, \code{\link{RBloomberg}}
}
\author{
Thomas P. Harte
}
\keyword{\code{\link{sort}}}
