% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trades.R
\name{getAllTrades}
\alias{getAllTrades}
\title{getAllTrades function}
\usage{
getAllTrades()
}
\description{
This function work only for IBKR accounts not for Gonet account
This function is used by other Tutils functions - NOT to be exported
No argument - takes its source from config::get()
}
