% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{getBSOptVega}
\alias{getBSOptVega}
\title{getBSOptVega}
\usage{
getBSOptVega(
  type,
  S,
  K,
  DTE,
  sig,
  r = as.numeric(config::get("interest_rate")),
  div = 0
)
}
\arguments{
\item{type}{a string, if equal to "Put" then getBSPutPrice is called, if equal to "Call" getBSCallPrice is called, if equal to "Stock" returns 1, else returns NA}

\item{S}{Spot price or current underlying price}

\item{K}{Strike price}

\item{DTE}{Number of days to expiration (may be a decimal number - useful as expiration may happen at 4:00pm or in the morning (index case))}

\item{sig}{annualized volatility}

\item{r}{Interest rate - default value is defined in config.yml file}

\item{div}{annualized dividend yield (i.e. sum of present values of upcoming dividends over 1 year divided by current spot price). Default value is 0.}
}
\description{
This function supplies vega value of an option based upon Black-Scholes-Merton model
}
\details{
It performs the computation as per textbooks using derivmkts package:

This function can be vectorized. It handles also special cases not handled correctly by bsput/bscall, such as:
\itemize{
\item DTE = 0 -> Vega = 0
\item DTE < 0 -> Vega = 0
}
}
\examples{
getBSOptVega(type="Call",S=100,K=100,r=0.04,DTE=5,sig=0.25,div=0)
getBSOptVega(type="Put",S=220,K=225,r=0.04,DTE=500,sig=0.3,div=0)
getBSOptVega(type="Put",S=220,K=247,r=0.04,DTE=700,sig=0.2,div=0)
getBSOptVega(type="Put",S=22,K=22.5,r=0.04,DTE=500,sig=0.3,div=0)
}
\keyword{Black-Scholes}
\keyword{trading}
