% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{getDTE}
\alias{getDTE}
\title{getDTE}
\usage{
getDTE(c_datetime, expdate)
}
\arguments{
\item{c_datetime}{is the current date&time}

\item{expdate}{is the expiration date (at 4:00pm)}
}
\description{
Given a list of numbers or dates,
The aim is to find the nearest within the list
The list does not need to be ordered
This function will proceed by looking at all list elements
}
\examples{
getDTE(as.Date("2023-12-10"),as.Date("2023-12-15"))
}
\keyword{Black-Scholes}
\keyword{date}
\keyword{options}
