% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{getImpliedVolStraddle}
\alias{getImpliedVolStraddle}
\title{getImpliedVolStraddle}
\usage{
getImpliedVolStraddle(
  S,
  K,
  r = as.numeric(config::get("interest_rate")),
  DTE,
  div = 0,
  price
)
}
\arguments{
\item{S}{Spot price or current underlying price}

\item{K}{Strike price}

\item{r}{Interest rate - default value is defined in config.yml file}

\item{DTE}{Number of days to expiration (may be a decimal number - useful as expiration may happen at 4:00pm or in the morning (index case))}

\item{div}{annualized dividend yield (i.e. sum of present values of upcoming dividends over 1 year divided by current spot price). Default value is 0.}

\item{price}{straddle price}
}
\description{
This function computes the implied volatility of a straddle, based upon Black-Scholes-Merton model
}
\details{
It performs the computation using \code{getImpliedVolOpt} function. As 30 days at-the-money option, this can be utilized as the standard implied volatility of the underlying.

This function cannot be vectorized.
}
\examples{
getImpliedVolStraddle(S=100,K=100,r=0.04,DTE=30,price=3.85)
getImpliedVolStraddle(S=22,K=22.5,r=0.04,DTE=30,price=1.1)
}
\keyword{Black-Scholes}
\keyword{trading}
\keyword{volatility}
