% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prices.R
\name{getLastAdjustedPrice}
\alias{getLastAdjustedPrice}
\title{getLastAdjustedPrice}
\usage{
getLastAdjustedPrice(ticker)
}
\arguments{
\item{ticker}{ticker name, as known by IBKR - can be one name or a vector of names}
}
\value{
a list of values (rounded to 2 decimals) corresponding to last values of tickers
}
\description{
This function takes one ticker (or a vector of tickers) as input and returns the last available adjusted value from Yahoo service.
}
\details{
To achieve this, it will call \code{getSymFromDate} with "today - 5" date - so to be sure to get at least one valid date,
even if there are week ends and closed days. It takes last open date if current date provided does not work.
See also \code{getLastPriceDate}
}
\examples{
getLastAdjustedPrice("SPY")
}
