% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prices.R
\name{getLastTickerData}
\alias{getLastTickerData}
\title{getLastTickerData}
\usage{
getLastTickerData(ticker)
}
\arguments{
\item{ticker}{string - IBKR style of ticker, if equal to "STOCK" or "All" then returns a list of NA values}
}
\value{
a list of 2 fields: \code{last} which contains last value of the ticker,
and \code{change} that contains a string giving the percentage of change since previous day.
}
\description{
For a given ticker this function returns the last known closed value (adjusted) and its change from previous day.
}
\details{
This function is not vectorized and accepts only one ticker at a time.
It calls Yahoo service through \code{getSymFromDate} to obtains necessary value.
}
\examples{
getLastTickerData("SPY")
}
