% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trades.R
\name{getOpenDate}
\alias{getOpenDate}
\title{getOpenDate}
\usage{
getOpenDate(trade_nr)
}
\arguments{
\item{trade_nr}{an integer or a vector of integers}
}
\value{
a data frame giving for each trade number an expiration date \code{expdate} and the original trade date \code{orig_date}
}
\description{
This function retrieves one or a vector of dates, all pertaining to open/adjusted trades
}
\details{
It works by matching all trade number in Trades.csv file argument against trade dates
Then the oldest date per trade is returned.
}
