% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prices.R
\name{getPriceAllDates}
\alias{getPriceAllDates}
\title{getPriceAllDates
This function is a helper function that helps to deal with \code{getSymFromDate} returned values
It extracts the first list of the list of lists returned by \code{getSymFromDate},
Takes the 6th column (adjusted values) and then convert it into a tibble with one column \code{date} and one column \code{value}.}
\usage{
getPriceAllDates(sym_list)
}
\arguments{
\item{sym_list}{a list of list, each list having 6 fields.}
}
\value{
a tibble with one column \code{date} and one column \code{value}.
}
\description{
getPriceAllDates
This function is a helper function that helps to deal with \code{getSymFromDate} returned values
It extracts the first list of the list of lists returned by \code{getSymFromDate},
Takes the 6th column (adjusted values) and then convert it into a tibble with one column \code{date} and one column \code{value}.
}
\examples{
getPriceAllDates(getSym("SPY"))
}
