% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prices.R
\name{getSym}
\alias{getSym}
\title{getSym}
\usage{
getSym(sym)
}
\arguments{
\item{sym}{one or a vector of symbols}
}
\value{
a list of list: each list contains the list of prices and volume: Open, Close, Low, High, Volume and Adjusted
}
\description{
This function retrieves all values from Yahoo starting from CurrentTradesInitialDate - see config.yml file at user level directory
This function calls \code{getSymFromDate} with one or a vector of tickers.
}
\examples{
getSym("SPY")
}
