% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prices.R
\name{getSymFromDate}
\alias{getSymFromDate}
\title{getSymFromDate}
\usage{
getSymFromDate(sym, date)
}
\arguments{
\item{sym}{one or a vector of symbols}

\item{date}{a start date from which to retrieve symbols}
}
\value{
a list of list: each list contains the list of prices and volume: Open, Close, Low, High, Volume and Adjusted
}
\description{
This function gets from Yahoo service all values (Open, High, Low, Close, Volume and Adjusted)
for one of a vector of symbols. It is based upon quantmod getSymbols function.
It converts IBKR-style tickers into Yahoo-style tickers first.
}
\examples{
getSymFromDate("SPY",as.Date("2023-12-01"))
}
\keyword{Yahoo}
