% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-utils.R
\name{\%not.in\%}
\alias{\%not.in\%}
\title{Negates \code{\link{\%in\%}}}
\usage{
... \%not.in\% NA
}
\description{
Negates \code{\link{\%in\%}}
}
\examples{
  assert("a" \%not.in\% letters[2:length(letters)])
  "a" \%not.in\% letters[2:length(letters)]
  "a" \%not.in\% letters

}
\seealso{
\code{\link{\%in\%}}, \code{\link{Negate}}
}
\author{
Thomas P. Harte
}
\keyword{\code{\link{Negate}}}
\keyword{\code{\link{\%in\%}},}
