% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/not_in.R
\name{\%not_in\%}
\alias{\%not_in\%}
\title{Not In}
\usage{
a \%not_in\% b
}
\arguments{
\item{a}{vector. values to search for}

\item{b}{vector. values to search within}
}
\value{
logical TRUE or FALSE
}
\description{
The inverse of \%in\%
}
