% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-utils.R
\name{grep_system}
\alias{grep_system}
\title{Use system \code{grep} (*NIX only)}
\usage{
grep_system(pattern, filename, options = "")
}
\arguments{
\item{pattern}{\code{\link{character}} to search}

\item{filename}{\code{\link{character}} file name}

\item{options}{\code{\link{character}} passed to \code{grep}}
}
\value{
\code{\link{character}} \code{\link{vector}} output from \code{cut}
}
\description{
Use system \code{grep} (*NIX only). For large files this is much faster
than reading the file into R and grepping the columns.
}
\examples{
tab<- read.csv(text='Name,Age,Salary,ID
            Dick,38,32k,1
            Tom,21,21k,2
            Harry,56,NA,3',
            header=TRUE,
            stringsAsFactors=FALSE
   )
   for (col in which(col_classes(tab)=="character"))
       tab[, col]<- tutils::trim(tab[, col])
   tab

   filename<- paste(tempfile(), ".csv", sep="")
   write.csv(tab, file=filename, row.names=FALSE, quote=FALSE)
   grep_system("Tom", filename)
   grep_system("Dick", filename)
   grep_system("harry", filename, options="-i")

   unlink(filename)

}
\seealso{
\code{\link{system}}, \code{\link{cut}}
}
\author{
Thomas P. Harte
}
\keyword{\code{\link{cut}}}
\keyword{\code{\link{system}},}
