% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object-utils.R
\name{ht}
\alias{ht}
\title{Returns the head and tail part of an object}
\usage{
ht(x, n = 6, ...)
}
\arguments{
\item{x}{an object.}

\item{n}{\code{\link{integer}}
If positive, size for the resulting object:
   number of elements for a vector (including lists), rows for a
   matrix or data frame or lines for a function. If negative,
   all but the 'n' last/first number of elements of 'x'}

\item{...}{arguments to \code{\link{head}} and \code{\link{tail}}}
}
\value{
\code{\link{character}}
}
\description{
Returns the head and tail part of an object
}
\examples{
 N<- 20
 df<- data.frame(Name=letters[1:N], Value=1:N)
 ht(df)
 N<- 3
 df<- data.frame(Name=letters[1:N], Value=1:N)
 ht(df)

}
\seealso{
\code{\link{head}}, \code{\link{tail}}
}
\author{
Thomas P. Harte
}
\keyword{\code{\link{sort}}}
