% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string-utils.R
\name{is_blank}
\alias{is_blank}
\title{checks if a string is blank}
\usage{
is_blank(x)
}
\arguments{
\item{x}{\code{\link{character}} string to check}
}
\value{
\code{\link{logical}}
}
\description{
checks if a string is blank
}
\examples{
  is_blank("")
  is_blank(rep("",4))
  is_blank(c("hello",rep("",4)))

}
\seealso{
\code{\link{character}}
}
\author{
Thomas P. Harte
}
\keyword{\code{\link{character}}}
