% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts-utils.R
\name{is_Date}
\alias{is_Date}
\title{Check if the object is of class \code{\link{Date}}}
\usage{
is_Date(x)
}
\arguments{
\item{x}{object}
}
\value{
\code{\link{logical}}
}
\description{
Check if the object is of class \code{\link{Date}}
}
\examples{
  is_Date(as.Date("2017-07-04"))
  is_Date("2017-07-04")

  x<- seq(as.Date("2017-07-01"), len=4, by=1)
  is_Date(x)
  is_Date(as.character(x))

}
\seealso{
\code{\link{Date}}
}
\author{
Thomas P. Harte
}
\keyword{\code{\link{Date}}}
