% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-utils.R
\name{is_even}
\alias{is_even}
\alias{is_odd}
\title{Checks for even / odd number}
\usage{
is_even(x)

is_odd(x)
}
\description{
Checks for even / odd number
}
\examples{
  is_even(1:10)
  is_odd(1:10)

}
\seealso{
\code{\link{\%\%}}
}
\author{
Thomas P. Harte
}
\keyword{\code{\link{\%\%}}}
