% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string-utils.R
\name{is_not_applicable}
\alias{is_not_applicable}
\title{Finds strings matching N/A}
\usage{
is_not_applicable(x)
}
\arguments{
\item{x}{\code{\link{character}} \code{\link{vector}}, or \code{\link{data.frame}}}
}
\value{
\code{\link{data.frame}} with same structure as \code{x}, but with
  entries of mode \code{\link{logical}}, or \code{\link{character}}
  \code{\link{vector}} with entries of mode \code{\link{logical}}
}
\description{
Finds strings matching N/A
}
\examples{
   tab<- read.table(text='Name| Age|Salary|ID
            Tom| NA|32k|1
            N/A| NA|21k|2
            Harry| NA|NA|N/A',
           header=TRUE,
           sep="|",
           colClasses=c("character","integer","character","character")
       )
   tab
   col_classes(tab)

   res<- is_not_applicable(tab)
   col_classes(res)
   res

}
\seealso{
\code{\link{NA}}
}
\author{
Thomas P. Harte
}
\keyword{\code{\link{NA}}}
