% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object-utils.R
\name{is_try_error}
\alias{is_try_error}
\title{Check for try-error}
\usage{
is_try_error(x)
}
\arguments{
\item{x}{\code{\link{try}}}
}
\value{
\code{\link{logical}}
}
\description{
Checks to see if a try-error occurred
}
\examples{
e<- new.env()
local({
	a<- "a"
	b<- "this is 'b'"
	x<- pi
}, env=e)
f<- new.env()
is_try_error(try(
	put_var("someVar", from=e, to=f)
))
whos(sort="Name", env=f)
# Error in get(var, envir = fromEnvir) : object 'someVar' not found
# [1] TRUE
#     Class Dimensions Bytes
#NULL    NA         NA    NA

}
\seealso{
\code{\link{try}}
}
\author{
Thomas P. Harte
}
\keyword{\code{\link{environment}}}
