% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string-utils.R
\name{is_whitespace}
\alias{is_whitespace}
\title{Finds strings that are purely whitespace}
\usage{
is_whitespace(x)
}
\arguments{
\item{x}{\code{\link{character}} \code{\link{vector}}, or \code{\link{data.frame}}}
}
\value{
\code{\link{data.frame}} with same structure as \code{x}, but with
  entries of mode \code{\link{logical}}, or \code{\link{character}}
  \code{\link{vector}} with entries of mode \code{\link{logical}}
}
\description{
Finds strings that are purely whitespace (note: not strings that \emph{contain} whitespace)
}
\examples{
   tab<- read.table(text='Name| Age|Salary|ID
            | |32k|1
            N/A| NA|21k|2
            Harry| NA|NA|NA',
           header=TRUE,
           sep="|",
           colClasses=c("character","character","character","integer")
       )
   tab
   col_classes(tab)
   # these two columns contain whitespace:
   tab[, 1]
   tab[, 2]

   res<- is_whitespace(tab)
   col_classes(res)
   res

}
\seealso{
\code{\link{gsub}}
}
\author{
Thomas P. Harte
}
\keyword{\code{\link{gsub}}}
