% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/businessDates.R
\name{isBusinessDay}
\alias{isBusinessDay}
\title{Predicate for business days}
\usage{
isBusinessDay(dates)
}
\arguments{
\item{dates}{Vector of Date objects}
}
\value{
Logical vector, same length as the input
}
\description{
Test a vector of dates for being business days.
Returns a vector of logical values, same length as the input:
TRUE if the date is a business day, FALSE if not.
}
