% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/journal.R
\name{journalApp}
\alias{journalApp}
\title{Journal app}
\usage{
journalApp()
}
\description{
This complete app makes use of journal UI/server module.
}
\details{
It can be used as a standalone application to record new entries or to read journal.
}
