% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/journal.R
\name{journalServer}
\alias{journalServer}
\title{Journal module server}
\usage{
journalServer(id, sym, windowDate = shiny::reactive(lubridate::today() - 365))
}
\arguments{
\item{id}{this is used by caller to identify journal and have the link with server piece}

\item{sym}{this is to filter only log entries corresponding to \code{sym}}

\item{windowDate}{this is to retrieve only log entries whose date is greater or equal to \code{windowdate}}
}
\description{
This module provides the ability to write log entries into a journal and to retrieve them
}
