% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/journal.R
\name{journalUI}
\alias{journalUI}
\title{Journal module UI}
\usage{
journalUI(id)
}
\arguments{
\item{id}{this is used by caller to identify journal and have the link with server piece}
}
\description{
This module provides the ability to write log entries into a journal and to retrieve them
}
