% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orientation.R
\name{landscape}
\alias{landscape}
\title{Orient a data structure into landscape}
\usage{
landscape(x, ...)
}
\arguments{
\item{x}{A matrix or data frame}
}
\value{
The input (\code{x}) in landscape orientation
}
\description{
This is handy for forcing a matrix or data frame
 into landscape orientation prior to printing.
}
