% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posts.R
\name{latestAlerts}
\alias{latestAlerts}
\title{Most-recent events that are alerts}
\usage{
latestAlerts(event_types = NULL)
}
\arguments{
\item{event_types}{If specified, limit to events of these types
(optional, character vector)}
}
\description{
Returns most-recent events that have \code{Alert} set to \code{TRUE}.
}
\seealso{
\link{latestEvents}, \link{postEvent}
}
