% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posts.R
\name{latestEvents}
\alias{latestEvents}
\title{Read the latest event messages.}
\usage{
latestEvents(event_types = NULL)
}
\arguments{
\item{event_types}{Limit to these event types;
NULL means all events (NULL or character)}
}
\description{
Read the latest event messages.
}
\seealso{
\link{loadEvents} describes the return type.
}
