% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posts.R
\name{latestStatus}
\alias{latestStatus}
\title{Most-recent status events}
\usage{
latestStatus()
}
\value{
Returns a data frame with columns
\itemize{
\item Origin
\item Timestamp
\item Status
\item Message
\item Alert
}
}
\description{
Most-recent status events
}
\seealso{
\link{postStatus}
}
