% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ligne.R
\name{ligneServer}
\alias{ligneServer}
\title{Journal ligne server}
\usage{
ligneServer(
  id,
  sym,
  mul,
  s_datetime,
  i_rate,
  u_price,
  div,
  p_days,
  p_u_price,
  currency,
  exchange,
  tradingClass,
  exp_dates,
  strikes
)
}
\arguments{
\item{id}{this is used by caller to identify line and have the link with server piece}

\item{sym}{this is to filter only log entries corresponding to \code{sym}}

\item{mul}{multiplier - usually 100, but also 10 (ESTX50)}

\item{s_datetime}{start date time- currently now}

\item{i_rate}{value for interest rate (options BS computation)}

\item{u_price}{value for underlying spot price}

\item{div}{dividend yield used by BS option computation}

\item{p_days}{number of days till projected date}

\item{p_u_price}{projected underlying price}

\item{currency}{string - "USD", "EUR" or "CHF"}

\item{exchange}{string usually "SMART", may be "CBOE" or "EUREX"}

\item{tradingClass}{string necessary to get option price from IBKR}

\item{exp_dates}{a vector of dates}

\item{strikes}{a vector of strikes}
}
\description{
This module provides the ability to write log entries into a journal and to retrieve them
}
