% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object-utils.R
\name{load_as}
\alias{load_as}
\title{Reload a single object written with the function save}
\usage{
load_as(file)
}
\arguments{
\item{file}{\code{\link{character}} file name of the stored object to reload}
}
\value{
\code{\link{character}}
}
\description{
\code{load_as} reloads the contents of a single object that was saved in .Rdata format into the workspace
}
\examples{
file<- paste(tempfile(), ".Rdata", sep="")
objectWithAVeryStrangeName<- rnorm(10)
whos()
save(objectWithAVeryStrangeName, file=file)
## not run:
## load(file)
## loads objWithAVeryStrangeName into the workspace which may be undesirable
x<- load_as(file)
whos()
unlink(file)

}
\seealso{
\code{\link{load}}, \code{\link{save}}
}
\author{
Thomas P. Harte
}
\keyword{\code{\link{sort}}}
