% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posts.R
\name{loadEvents}
\alias{loadEvents}
\title{Read the entire events file}
\usage{
loadEvents(event_types = NULL)
}
\value{
Returns a data frame with columns
\itemize{
\item Origin (character)
\item Timestamp (POSIXct)
\item EventType (character)
\item Payload (list-col)
\item Alert (logical)
}
}
\description{
Read the entire events file
}
\seealso{
This is a lower-level function.
You probably want to call \link{latestEvents} or \link{todaysEvents} instead.
}
