% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object-utils.R
\name{match_col}
\alias{match_col}
\title{Find row-col location of a search string in \code{\link{data.frame}}}
\usage{
match_col(term, DF, ROW.FUN = c("first", "last"), COL.FUN = c("first",
  "last"))
}
\arguments{
\item{term}{\code{\link{character}} string to search for}

\item{DF}{\code{\link{data.frame}} or \code{\link{matrix}} to search}

\item{ROW.FUN}{\code{\link{function}} for row search}

\item{COL.FUN}{\code{\link{function}} for column search}
}
\value{
named \code{\link{vector}} with row & col location of search
}
\description{
Description: What the function does in more detail
}
\examples{
tab<- read.csv(con<- textConnection(
"Name,    Age, Salary
 Derek  ,  NA,    32k		# <- NOTE: 'Derek  '
 Tom,      26,    21k
 NA,       NA,     NA
 Harry,    31,    50k"
), header=TRUE, colClasses=c("character","integer","character"), comment.char="#"); close(con)

match_col("NON-MATCHING-STRING", tab, ROW.FUN="first", COL.FUN="first")
match_col("NON-MATCHING-STRING", tab, ROW.FUN="first", COL.FUN="last")
match_col("NON-MATCHING-STRING", tab, ROW.FUN="last", COL.FUN="first")
match_col("NON-MATCHING-STRING", tab, ROW.FUN="last", COL.FUN="last")

result<-        rep(NA,2)
names(result)<- c("row","col")

result["row"]<- 1; result["col"]<- 1
all.equal(match_col("k", tab, ROW.FUN=tutils::first, COL.FUN=tutils::first), result)

result["row"]<- 1; result["col"]<- 3
all.equal(match_col("k", tab, ROW.FUN=tutils::first, COL.FUN=tutils::last), result)

result["row"]<- 4; result["col"]<- 1
all.equal(match_col("r", tab, ROW.FUN=tutils::last, COL.FUN=tutils::first), result)	# <- NOTE: testing for "r"

result["row"]<- 4; result["col"]<- 3
all.equal(match_col("k", tab, ROW.FUN=tutils::last, COL.FUN=tutils::last), result)	# <- NOTE: testing for "k"

}
\seealso{
\code{\link{match}}, \code{\link{grep}}
}
\author{
Thomas P. Harte
}
\keyword{\code{\link{grep}}}
\keyword{\code{\link{match}},}
