% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantileCI.R
\name{medianCI}
\alias{medianCI}
\title{Estimate and confidence interval of the median}
\usage{
medianCI(y, conf.level = 0.95, method = "binomial", na.rm = FALSE)
}
\arguments{
\item{y}{Vector of sample data}

\item{method}{Method}

\item{na.rm}{If TRUE, remove NA values from data}

\item{alpha}{Tail probability}
}
\value{
A named, three-value vector: estimate, lowerCI, upperCI
}
\description{
Estimate and confidence interval of the median
}
