% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthFunctions.R
\name{monthCode}
\alias{monthCode}
\title{Return a futures month code}
\usage{
monthCode(m)
}
\arguments{
\item{m}{A month number, month name, Date, or yearmon}
}
\value{
One-character futures month code,
F = Jan, G = Feb, H = March, etc.
}
\description{
Return a futures month code
}
