% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datePartFactors.R, R/monthFunctions.R
\name{monthFactor}
\alias{monthFactor}
\alias{monthFactor.Date}
\alias{monthFactor.zoo}
\alias{monthFactor.character}
\alias{monthFactor.yearmon}
\title{Create month-of-year factor from zoo or xts object}
\usage{
monthFactor(x, ...)

monthFactor(x, ...)

\method{monthFactor}{Date}(d)

\method{monthFactor}{zoo}(z)

\method{monthFactor}{character}(ch)

\method{monthFactor}{yearmon}(ym)
}
\arguments{
\item{z}{The zoo or xts object}
}
\value{
A factor

Factor with levels Jan, Feb, Mar, ...
}
\description{
Given a zoo or xts object, return a month factor
based on the index. Factor levels are Jan, Feb, Mar, ...
}
\details{
TODO: Really, there should be monthFactor.zoo,
monthFactor.xts, monthFactor.Date, monthFactor.character, monthFactor.integer, etc.
}
\seealso{
months
}
