% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthFunctions.R
\name{monthIndex}
\alias{monthIndex}
\alias{monthIndex.Date}
\alias{monthIndex.POSIXlt}
\alias{monthIndex.integer}
\alias{monthIndex.numeric}
\alias{monthIndexToDate}
\title{Serial number of a month}
\usage{
monthIndex(x, ...)

\method{monthIndex}{Date}(x)

\method{monthIndex}{POSIXlt}(x)

\method{monthIndex}{integer}(year, month)

\method{monthIndex}{numeric}(year, month)

monthIndexToDate(mi)
}
\arguments{
\item{x}{Method dependent}
}
\value{
A serial integer for the month
}
\description{
Given a date or a year/month combination,
 return the number of months since Jan 1, 1970.
}
