% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthFunctions.R
\name{monthNumber}
\alias{monthNumber}
\alias{monthNumber.default}
\alias{monthNumber.integer}
\alias{monthNumber.numeric}
\alias{monthNumber.character}
\alias{monthNumber.Date}
\alias{monthNumber.yearmon}
\title{Convert month in character format to integer (1, ..., 12)}
\usage{
monthNumber(x, ...)

\method{monthNumber}{default}(x)

\method{monthNumber}{integer}(x)

\method{monthNumber}{numeric}(x)

\method{monthNumber}{character}(s)

\method{monthNumber}{Date}(d)

\method{monthNumber}{yearmon}(d)
}
\arguments{
\item{x}{Can be a number (1, ..., 12), name ("jan", "feb", ...),
futures month code ("F", "G", "H", ...),
Date, or yearmon.
Case is ignored for character arguments.}
}
\value{
Month number (1L, ..., 12L)
}
\description{
Convert month in character format to integer (1, ..., 12)
}
