% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/businessDates.R
\name{nearestBusinessDay}
\alias{nearestBusinessDay}
\title{Find business date nearest to given date}
\usage{
nearestBusinessDay(dates = Sys.Date())
}
\arguments{
\item{dates}{Vector of Date objects}
}
\value{
Vector of Date objects
}
\description{
Find business date nearest to given date
}
