% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{non.na}
\alias{non.na}
\title{Test for non-NA}
\usage{
non.na(x)
}
\arguments{
\item{x}{Any R object}
}
\value{
TRUE if \code{x} is not NA, FALSE otherwise.
}
\description{
Return TRUE if argument is not NA, FALSE otherwise
}
