% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{non.null}
\alias{non.null}
\title{Test for non-NULL}
\usage{
non.null(x)
}
\arguments{
\item{x}{Any R object}
}
\value{
TRUE if \code{x} is not NULL, FALSE otherwise.
}
\description{
Return TRUE if argument is not NULL, FALSE otherwise
}
