% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nth}
\alias{nth}
\title{Return Nth item of a vector or Nth row of a 2-dimensional structure}
\usage{
nth(x, n)
}
\arguments{
\item{x}{Vector, matrix, or data frame}

\item{n}{Index of item or row}
}
\value{
Nth item of vector, or Nth row of matrix or data frame
}
\description{
Return Nth item of a vector or Nth row of a 2-dimensional structure
}
