% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/on_off.R
\name{on_off}
\alias{on_off}
\title{On or off?}
\usage{
on_off(b)
}
\arguments{
\item{b}{A boolean value}
}
\value{
Returns "on" if b is a single, defined boolean value that is TRUE;
"off" otherwise
}
\description{
On or off?
}
