% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{op-disjunction-functional}
\alias{op-disjunction-functional}
\alias{\%or\%}
\title{Combine two predicates}
\usage{
p1 \%or\% p2
}
\arguments{
\item{p1}{A predicate of one argument}

\item{p2}{A predicate of one argument}
}
\value{
Returns a \emph{function}, not a value. (Be careful.)
}
\description{
Given two predicates, this operator combines them
 into one, new predicate. The new predicate calls
 the given predicates, applying "||" to the results.
}
\details{
Typical usage within the Tau software would be
 \code{decl(x, is.null \%or\% is.numeric)}.
}
\examples{
(is.null \%or\% is.numeric)(NULL)
(is.null \%or\% is.numeric)(pi)
(is.null \%or\% is.numeric)("foo")

}
