% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object-utils.R
\name{order_cn}
\alias{order_cn}
\alias{order_cn_front}
\alias{order_cn_back}
\title{Re-order named columns to front or back of \code{\link{data.frame}}}
\usage{
order_cn_front(x, cn)

order_cn_back(x, cn)
}
\arguments{
\item{x}{\code{\link{data.frame}}}

\item{cn}{\code{\link{character}} \code{\link{vector}} of column names}
}
\value{
\code{\link{data.frame}} with re-ordered columns
}
\description{
Re-order named columns to front or back of \code{\link{data.frame}}
}
\examples{
tab<- read.table(
       con<- textConnection(
           "Name    Age Salary ID
            Dick     38    32k  1
            Tom      21    21k  2
            Harry    56     NA  3"
       ),
       header=TRUE,
       colClasses=c("character","integer","character","integer")
   )
   close(con)

 tab \%>\% order_cn_front(c("Salary","Age"))
 tab \%>\% order_cn_back(c("Salary","Age"))

}
\seealso{
\code{\link{data.frame}}, \code{\link{dplyr::everything}},
  \url{http://stackoverflow.com/questions/27865865/in-dplyr-how-to-delete-and-rename-columns-that-dont-exist-manipulate-all-name}
}
\author{
Thomas P. Harte
}
\keyword{\code{\link{data.frame}},}
\keyword{\code{\link{dplyr::everything}}}
