% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partialCor.R
\name{partialCor}
\alias{partialCor}
\title{Calculate partial correlation}
\usage{
partialCor(x, y, z, use = "complete.obs")
}
\arguments{
\item{x}{A vector}

\item{y}{A vector}

\item{z}{A vector}

\item{use}{Passed to 'use' parameter of cor()}
}
\value{
Partial correlation
}
\description{
Note that all input vectors must be the same length.
}
