% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-utils.R
\name{pdftikz}
\alias{pdftikz}
\title{Use system \code{pdflatex} (*NIX only) to compile tikz output}
\usage{
pdftikz(filename)
}
\arguments{
\item{filename}{\code{\link{character}} file name}
}
\value{
output of \code{tryCatch} block
}
\description{
Use system \code{pdflatex} (*NIX only) to compile tikz output. 
\code{\link{pdftikz}} will save the current directory, change
to the directory of the filename and execute pdflatex on the 
tikz output to produce a pdf file
}
\examples{
# reproduce cover PDF from:
#     https://cran.r-project.org/web/packages/tikzDevice/vignettes/tikzDevice.pdf
d<- tempdir()
filename<- file.path(d, 'cover-pdf.tex')
require(tikzDevice)
tikz(file=filename, width=4, height=4, standAlone=TRUE)
par(las=1)

x <- seq(-4.5,4.5,length.out=100)
y <- dnorm(x)

xi <- seq(-2,2,length.out=30)
yi <- dnorm(xi)

plot(x,y,type='l',col='blue',ylab='$p(x)$',xlab='$x$')
lines(xi,yi,type='s')
lines(range(xi),c(0,0))
lines(xi,yi,type='h')
title(main="$p(x)=\\\\frac{1}{\\\\sqrt{2\\\\pi}}e^{-\\\\frac{x^2}{2}}$")
int <- integrate(dnorm,min(xi),max(xi),subdivisions=length(xi))
text(2.8, 0.35,
    paste(
	"\\\\small$\\\\displaystyle\\\\int_{",min(xi),"}^{",max(xi),"}p(x)dx",
	"\\\\approx", round(int[['value']], 3), '$', 
        sep=''
    )
)

dev.off()
pdftikz(filename)

}
\seealso{
\code{\link{system}}
}
\author{
Thomas P. Harte
}
\keyword{\code{\link{system}},}
\keyword{\code{pdflatex}}
