% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{pgcd}
\alias{pgcd}
\title{pgcd}
\usage{
pgcd(x)
}
\arguments{
\item{x}{is a vector of integers, may include NA}
}
\description{
"plus grand commun diviseur"
GCD function from DescTools has several limitations that are addressed in this package:
\itemize{
\item when there is only one value, GCD raises an error
\item GCD does not know how to handle NA
}
}
\examples{
pgcd(c(6,4))  ### Should be equal to 2
pgcd(c(6,4,NA))  ### Should be equal to 2
pgcd(5) ### Equal to 5
}
\keyword{GCD}
\keyword{arithmetic}
