% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{Polarity}
\alias{Polarity}
\alias{is.Polarity}
\alias{as.Polarity}
\title{Polarity type}
\usage{
Polarity(x)

is.Polarity(x)

as.Polarity(x)
}
\arguments{
\item{x}{A vector integer values +1 or -1}
}
\description{
The \code{Polarity} type indicates a \emph{long} or \emph{short} position
by using +1 and -1.
\itemize{
\item +1 indicates a long position.
\item -1 indicates a short position
}

Note that this indicates only the polarity (long vs. short),
not the size of the position.
}
\seealso{
\link{Side} type, \link{Tone} type
}
