% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orientation.R
\name{portrait}
\alias{portrait}
\title{Force a data structure into portrait orientation}
\usage{
portrait(x, ...)
}
\arguments{
\item{x}{A matrix or data frame}
}
\value{
The input (\code{x}) in portrait orientation
}
\description{
This is handy for forcing a matrix or data frame
 into portrait orientation prior to printing.
}
