% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/status.R
\name{post_alert}
\alias{post_alert}
\title{Post an alert}
\usage{
post_alert(
  name,
  status = "error",
  message = NA_character_,
  url = NA_character_
)
}
\arguments{
\item{name}{(character)}

\item{status}{(character)}

\item{message}{(character or list of character))}

\item{url}{(character)}
}
\description{
Post the status of a process, and alert the user to the status change.

This is a convenience function that calls \link{post_status} with
\code{alert=TRUE}.
Parameters are same as to \link{post_status}.
}
\examples{
# Post an error status and ask for notification of user
post_alert("other-process", status = "error")

}
