% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posts.R
\name{postAlert}
\alias{postAlert}
\title{Post an alert}
\usage{
postAlert(
  origin,
  message,
  expiration = NULL,
  topic = NA_character_,
  url = NA_character_,
  link_text = NA_character_
)
}
\arguments{
\item{origin}{(character)}

\item{message}{(optional, character)}

\item{expiration}{Date or time at which alert expires,
default to next business day after the current trading day
(optional, Date or POSIXt)}

\item{topic}{(optional, character)}

\item{url}{(optional, character)}

\item{link_text}{(optional, character)}
}
\value{
Nothing
}
\description{
(TBD)
}
\seealso{
\link{readAlerts} to read all open alerts
}
