% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posts.R
\name{postEvent}
\alias{postEvent}
\title{Write one event to the event file}
\usage{
postEvent(origin, event_type, payload, alert = FALSE)
}
\arguments{
\item{origin}{(character)}

\item{event_type}{(character)}

\item{payload}{An R object which can be serialized into JSON format}

\item{alert}{If TRUE, this event will be brought to the attention
of the system user}
}
\description{
Write one event to the event file
}
\seealso{
To post a \emph{status} event, call \link{postStatus}.

To read events, try \link{latestEvents} and \link{todaysEvents}.
To read status messages, try \link{latestStatus}.
To read alert messages, try \link{latestAlerts} and \link{todaysAlerts}.
}
