% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posts.R
\name{postStatus}
\alias{postStatus}
\title{Post a status message to the event log}
\usage{
postStatus(origin, status, message = NULL, alert = FALSE)
}
\arguments{
\item{origin}{Name of program posting status (character)}

\item{status}{Typically "OK" or "Error" or "Failure" (character)}

\item{message}{Useful, descriptive text message for user
(optional, character)}

\item{alert}{If TRUE, alert user to this status (logical)}
}
\description{
Post a status message to the event log
}
\seealso{
See \link{latestStatus} for reading status messages.
See \link{postEvent} for writing non-status messages.
}
