% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prependClass.R
\name{prependClass}
\alias{prependClass}
\title{Prepend a class attribute to existing class attribute}
\usage{
prependClass(x, className)
}
\arguments{
\item{x}{An object}

\item{className}{Class name to prepend to existing class attribute of x}
}
\value{
Returns x with updated class attribute
}
\description{
Prepend a class attribute to existing class attribute
}
\examples{
prependClass(mtcars, "Foo") |> class()
}
