% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/businessDates.R
\name{prevBusinessDay}
\alias{prevBusinessDay}
\alias{nextBusinessDay}
\title{Previous or next business day}
\usage{
prevBusinessDay(dates)

nextBusinessDay(dates)
}
\arguments{
\item{dates}{Vector of Date objects}
}
\value{
\code{prevBusinessDay} returns the vector of
corresponding previous business days.

\code{nextBusinessDay} returns the vector of
corresponding next business days.
}
\description{
Previous or next business day
}
