% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/businessDates.R
\name{prevTradingDay}
\alias{prevTradingDay}
\title{Previous trading date}
\usage{
prevTradingDay()
}
\value{
Date object
}
\description{
Returns the trading date immediately prior to
the current trading date.
}
\seealso{
\link{thisTradingDay}, which returns the current trading date.
}
