% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj_library.R
\name{proj_library}
\alias{proj_library}
\title{Attach the files in project's R directory}
\usage{
proj_library(files = NULL, ...)
}
\arguments{
\item{files}{Vector or list of files to source;
NULL means all *.R files (character)}

\item{...}{Passed to \code{source}}
}
\value{
Nothing
}
\description{
This will source all *.R files in the R subdirectory
 of the current project.
 It loads their contents into a new environment,
 and attaches that environment at the top of the search
 path (but underneith the global environment).
}
\seealso{
\link{proj_source} for sourcing project files
 into the global environment,
 \link{detach} for removing the new environment
 from the search path.
}
