% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj_source.R
\name{proj_source}
\alias{proj_source}
\title{Source files in project's R directory}
\usage{
proj_source(files = NULL, ...)
}
\arguments{
\item{files}{Vector or list of files to source;
NULL means all *.R files (character)}

\item{...}{Passed to \code{source}}
}
\value{
Nothing
}
\description{
This will source all *.R files in the R subdirectory
of the current project.
Their contents are loaded into the global environment.
}
\seealso{
\link{proj_library} for sourcing project files
into a new, attached environment.
}
