% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object-utils.R
\name{push_back}
\alias{push_back}
\title{appends an object to the back of a list}
\usage{
push_back(x, y)
}
\arguments{
\item{x}{\code{\link{list}} list of objects}

\item{y}{\code{\link{list}} list of objects to append to \code{x}}
}
\value{
\code{\link{character}}
}
\description{
appends an object to the back of a list
}
\examples{
# zero-length list x: append zero-length object y:
push_back(list(), list())

# zero-length list x: append non-list object:
x<- list()
x<- push_back(x, "one")
x

# unnamed list x: append non-list object:
x<- list()
x[[1]]<- "one"
x<- push_back(x, 2)
x

# named list x: append non-list object:
x<- list()
x[["one"]]<- 1
x<- push_back(x, "two")
x

# unnamed list x: append unnamed list object y:
x<- list()
x[[1]]<- "one"
y<- list(); y[[1]]<- 2;
x<- push_back(x, y)
x

# unnamed list x: append named list object y:
x<- list()
x[[1]]<- "one"
y<- list(); y[[1]]<- 2; names(y)<- "two"
x<- push_back(x, y)
x

# unnamed list x: append named list object y:
x<- list()
x[[1]]<- 1; names(x)<- "one"
y<- list(); y[[1]]<- 2; names(y)<- "two"
x<- push_back(x, y)
x

# note that a particularly useful idiom is to use
# push_back to extract the named elements of a list:
push_back(list(), x)
# thus, if you wish to concatenate the *elements* of
# a list with another list without simply pushing
# back the entire list 'x':
push_back(y, x)
# then use this idiom thusly:
push_back(y, push_back(list(), x))

}
\seealso{
\code{\link{list}}
}
\author{
Thomas P. Harte
}
\keyword{\code{\link{gsub}}}
