% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object-utils.R
\name{put_var}
\alias{put_var}
\title{copy object}
\usage{
put_var(var, from.envir = parent.frame(), to.envir = .GlobalEnv)
}
\arguments{
\item{var}{\code{\link{character}} name of variable in \code{from.envir} to copy}

\item{from.envir}{\code{\link{environment}} where variable is to be copied from}

\item{to.envir}{\code{\link{environment}} where variable is to be copied to}
}
\value{
\code{\link{invisible}}
}
\description{
copy a single variable from one environment to another environment
}
\examples{
	 e<- new.env()
	 local({
		a<- "a"
		b<- "this is 'b'"
		x<- pi
	 }, env=e)
  f<- new.env()
	 put_var("x", from=e, to=f)
	 whos(sort="Name", env=f)
}
\seealso{
\code{\link{copy.vars}}
}
\author{
Thomas P. Harte
}
\keyword{\code{\link{environment}}}
