% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantileCI.R
\name{quantileCI}
\alias{quantileCI}
\title{Estimate and confidence intervals of a quantile}
\usage{
quantileCI(y, prob, conf.level = 0.95, method = "binomial", na.rm = FALSE, ...)
}
\arguments{
\item{y}{A vector of sample data}

\item{prob}{Numeric probability in range [0, 1]}

\item{alpha}{Tail probability}
}
\value{
Three-element,named vector: estimate, lowerCI, upperCI
}
\description{
Estimate and confidence intervals of a quantile
}
