% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datePartFactors.R
\name{quarterFactor}
\alias{quarterFactor}
\title{Create quarter-of-year factor from zoo or xts object

Returns a factor with four levels (Q1, Q2, Q3, Q4),
taken from index of the parameter.}
\usage{
quarterFactor(z)
}
\arguments{
\item{z}{A zoo or xts object}
}
\value{
A factor
}
\description{
Create quarter-of-year factor from zoo or xts object

Returns a factor with four levels (Q1, Q2, Q3, Q4),
taken from index of the parameter.
}
