% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object-utils.R
\name{rbind_pad}
\alias{rbind_pad}
\title{Alternative to \code{\link{plyr::rbind.fill}} which can be memory-intensive}
\usage{
rbind_pad(x, verbose = FALSE)
}
\arguments{
\item{x}{\code{\link{list}} of objects inheriting from \code{\link{data.frame}}}

\item{verbose}{\code{\link{logical}} print output information}
}
\value{
\code{\link{data.frame}}
}
\description{
A plausible alternative to \code{\link{plyr::rbind.fill}}  which can be very memory-intensive
}
\examples{
  x<- list(
       mtcars[c("mpg", "wt")],
       mtcars[c("wt", "cyl")]
  )

  rbind_pad(x)
  plyr::rbind.fill(x)

}
\seealso{
\code{\link{plyr::rbind.fill}}
}
\author{
Thomas P. Harte
}
\keyword{\code{\link{plyr::rbind.fill}}}
