% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/status.R
\name{read_current_status}
\alias{read_current_status}
\title{Read all current status files}
\usage{
read_current_status()
}
\value{
Returns a data frame of status information.
Returns an empty data frame if there are not current
status reports.
}
\description{
A status file is \emph{current} if its timestamp
is the current settlement date or later
}
