% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posts.R
\name{readAlerts}
\alias{readAlerts}
\title{Read all non-expired alerts}
\usage{
readAlerts(subset = c("recent", "all"))
}
\arguments{
\item{subset}{The rows to return. By default, only the most recent alerts,
for a given \code{Origin} and \code{Topic}, are returned.}
}
\value{
Returns a data frame with columns
\itemize{
\item Origin (character)
\item Timestamp (POSIXct)
\item Message (character)
\item Expiration (POSIXct)
\item Topic (character or NA)
\item URL (character or NA)
\item LinkText (character or)
}

The data frame will be empty if there are no active alerts.
}
\description{
Return all open alerts; that is, alerts which have not yet expired.
}
\seealso{
\link{postAlert} for posting an alert in the first place.
}
