% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account.R
\name{readPortfolio}
\alias{readPortfolio}
\title{readPortfolio}
\usage{
readPortfolio(accountnr)
}
\arguments{
\item{accountnr}{is the account number (IBKR)}
}
\value{
a tibble with the following columns:
\code{date; heure; type; symbol; expiration; strike; Put/Call; pos; mktPrice; optPrice}
\code{mktValue; avgCost; uPnL; IV; pvDividend; delta; gamma; vega; theta; uPrice; multiplier; currency}
}
\description{
This function reads the \emph{account_nr.csv} file and performs a bunch of data wrangling before returning a tibble with all data.
}
\details{
Data wrangling:
\enumerate{
\item formats it with right Date and HMS format
\item converts \code{position} to \code{integer}
\item removes all CASH positions
\item rename IBKR columns to more friendly names
}

It then filters data so that it matches \code{accountnr} number
}
\examples{
readPortfolio("D100000")

}
