% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posts.R
\name{readStatus}
\alias{readStatus}
\title{Read Tau status postings}
\usage{
readStatus(origin = NULL)
}
\arguments{
\item{origin}{Limit to postings from this origin (optional, character)}
}
\value{
Returns a data frame with columns
\itemize{
\item Origin (character)
\item Timestamp (POSIXct)
\item Status (character)
\item Message (character or NA)
\item URL (character or NA)
\item LinkText (character or NA)
}

Returns an empty data frame if no status postings are available.
}
\description{
blah blah blah
}
\seealso{
\link{postStatus} for posting a status in the first place
}
