% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{reformat}
\alias{reformat}
\title{reformat}
\usage{
reformat(file)
}
\arguments{
\item{file}{is the input CSV file}
}
\description{
reformat function is to be used to convert a CSV file
separated by commas into a file CSV separated by semi-columns
The new file has an extension csv2 and the same name
}
\keyword{csv}
\keyword{file}
\keyword{locale}
