% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object-utils.R
\name{remove_from}
\alias{remove_from}
\alias{remove_from_rows}
\alias{remove_from_cols}
\title{Remove all columns or rows from \code{\link{data.frame}} matching criterion}
\usage{
remove_from(DF, fun, dim = c("both", "rows", "cols"))

remove_from_rows(DF, fun)

remove_from_cols(DF, fun)
}
\arguments{
\item{DF}{\code{\link{data.frame}} or  \code{\link{matrix}}}

\item{fun}{\code{\link{function}} with criterion for removal}
}
\value{
\code{\link{data.frame}} or \code{\link{matrix}} with matching removed
}
\description{
Remove all columns or rows from \code{\link{data.frame}} matching criterion
}
\examples{
FUN<- function(x) is_blank(x) | is.na(x)

tab<- read.table(con<- textConnection(
"Name  Age Salary
 Dick   NA    32k
 Tom    NA    21k
 NA     NA    NA"
), header=TRUE, colClasses=c("character","integer","character")); close(con)


 remove_from_rows(tab, fun=FUN)
 remove_from(tab, fun=FUN, dim="row")

 remove_from_cols(tab, fun=FUN)
 remove_from(tab, fun=FUN, dim="col")

 remove_from(tab, fun=FUN, dim="both")

}
\seealso{
\code{\link{grep}}, \code{\link{match}}
}
\author{
Thomas P. Harte
}
\keyword{\code{\link{grep}},}
\keyword{\code{\link{match}}}
