% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorigin.R
\name{reorigin}
\alias{reorigin}
\title{Reorigin data to zero}
\usage{
reorigin(x, ...)
}
\arguments{
\item{x}{An R object}
}
\value{
Same object, but reorigined to zero
}
\description{
Translate data so that first element is zero
}
