% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safeAverage.R
\name{safeAverage}
\alias{safeAverage}
\title{Safely compute an average and it's confidence interval}
\usage{
safeAverage(x, average = "mean", conf.level = 0.95, na.rm = TRUE)
}
\arguments{
\item{x}{A numeric vector}

\item{average}{Either "mean" or "median"}

\item{conf.level}{Confidence level for confidence interval}

\item{na.rm}{If TRUE, remove NA values}
}
\value{
A 3-element vector: estimate, lowerCI, and upperCI
}
\description{
Safely compute an average and it's confidence interval
}
