% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object-utils.R
\name{save_zoo}
\alias{save_zoo}
\title{Save zoo object to named extension file}
\usage{
save_zoo(z, file = character(0), append = FALSE, ...)
}
\description{
Saves a \code{\link{zoo}} object to a file with a named extension
}
\details{
@param z \code{\link{zoo}}
	  a zoo object (\code{\link{coredata}} can be multi-column)
 @param file \code{\link{character}}
	  filename in which to save \code{z} with the file extension being one of:
        \itemize{
                \item{"psv"}{pipe-separated values}
                \item{"csv"}{comma-separated values}
                \item{"txt"}{tab-separated values}
                \item{"xls"}{Excel spreadsheet}
                \item{"Rdata"}{R binary file}
        }
}
\examples{
 require(zoo)
	n<- 10
	dates<- seq(as.Date("2008-10-31"), by=1, len=n)
	z<- zoo(matrix(c(rnorm(n, sd=.1), rnorm(n, sd=.5), rnorm(n, sd=1)), nc=3),
		order.by=dates)
	basename<- tempfile("z")

	# save as a PSV file:
	filename<- paste(basename, ".psv", sep="")
	save_zoo(z, file=filename)
	# note: column names added by default:
	file.show(filename)

	# add column names:
	colnames(z)<- c("one", "two", "three")
	# this time save as a CSV file:
	filename<- paste(basename, ".csv", sep="")
	save_zoo(z, file=filename)
	file.show(filename)

	# this time save as a TXT file:
	filename<- paste(basename, ".txt", sep="")
	save_zoo(z, file=filename)
	file.show(filename)

	filename<- paste(basename, ".Rdata", sep="")
	save_zoo(z, file=filename)
	# get rid of 'z':
	rm(z)
	# restore 'z':
	print(load(filename))
	# restore 'z' as 'foo':
	foo<- load_as(filename)
	stopifnot(identical(z, foo))

}
\seealso{
\code{\link{save}},
         \code{\link{write.table}},
	        \code{\link{write.csv}},
         \code{\link{write.xls}},
         \code{\link{get_file_ext}},
         \code{\link{zoo}}
}
\author{
Thomas P. Harte
}
\keyword{\code{\link{zoo}}}
