% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object-utils.R
\name{set_fns}
\alias{set_fns}
\alias{Intersect}
\alias{Union}
\title{Cumulative set functions.}
\usage{
Intersect(x)

Union(x)
}
\arguments{
\item{x}{\code{list} containing objects to which \code{\link{intersect}} can be applied}
}
\value{
intersection of \code{list} objects
}
\description{
Cumulative set functions.
}
\examples{
  sets<- list(
      1:3,
      2:4,
      3:5
  )

  unlist(last(Intersect(sets)))
  unlist(last(Union(sets)))

}
\seealso{
\code{\link{intersect}}, \code{\link{Reduce}}
}
\author{
Thomas P. Harte
}
\keyword{\code{\link{Reduce}}}
\keyword{\code{\link{intersect}},}
