% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{setColnames}
\alias{setColnames}
\title{Set column names in a pipe}
\usage{
setColnames(x, names)
}
\arguments{
\item{x}{Anything that has column names}

\item{names}{Character vector of new column names}
}
\description{
This is like assigning column names, but in a pipe-friendly way.
}
\seealso{
\link[magrittr:aliases]{magrittr::set_colnames}, from which this idea was stolen
}
