% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{Side}
\alias{Side}
\alias{is.Side}
\alias{as.Side}
\title{Side type}
\usage{
Side(x)

is.Side(x)

as.Side(x)
}
\arguments{
\item{x}{A character vector with values "long" and "short".}
}
\description{
The \code{Side} type indicates a \emph{long} or \emph{short} position
by using the strings "long" and "short".
\itemize{
\item "long" indicates position that is owned
\item "short" indicates a position sold short
}
}
\seealso{
\link{Polarity} type, \link{Tone} type
}
