% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signalError.R
\name{signalError}
\alias{signalError}
\alias{signalErrorIf}
\alias{signalErrorIfNot}
\title{Signal an error condition}
\usage{
signalError(..., sep = " ", log = TRUE)

signalErrorIf(b, ...)

signalErrorIfNot(b, ...)
}
\arguments{
\item{...}{Character strings that will be pasted together into an error message}

\item{sep}{Separator for strings in error message (character)}

\item{log}{If TRUE, log the error}
}
\value{
Returns nothing. Called for its effect.
}
\description{
Signal an error condition. Possibly log the error, too.
}
\seealso{
The message is logged via \code{\link[logger:log_level]{logger::log_error()}}.
}
