% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smartEllipsis.R
\name{smartEllipsis}
\alias{smartEllipsis}
\title{Join a list of strings, using ellipsis if necessary}
\usage{
smartEllipsis(strings, max = 2, tail = 1, empty = "None")
}
\arguments{
\item{strings}{Vector of strings to be joined (character)}

\item{max}{Number of strings before ellipsis (integer)}

\item{tail}{Number of strings after the ellipsis (integer, optional)}

\item{empty}{Returned if \code{strings} is empty (character)}
}
\value{
All the \code{strings} pasted together, separated by commas,
up to a maximum of \code{max} strings;
if \code{strings} has more than \code{max} elements,
the extra elements are replaced by an ellipsis.
If \code{strings} is empty, then \code{empty} is returned.
}
\description{
This joins a vector of strings into one string
with the elements separated by commas.
If there are too many inputs, however,
will use an ellipsis to shorten the output.
}
\examples{
# This returns "tom, dick, harry, peter"
smartEllipsis(c("tom", "dick", "harry", "peter"), 4)

# This also ret urns "tom, dick, harry, peter"
smartEllipsis(c("tom", "dick", "harry", "peter"), 3)

# This returns "tom, dick, harry, ..."
smartEllipsis(c("tom", "dick", "harry", "peter"), 3, tail = 0)

# This returns "tom, dick, ..., peter"
smartEllipsis(c("tom", "dick", "harry", "peter"), 2)

# This returns "tom, dick, ..."
smartEllipsis(c("tom", "dick", "harry", "peter"), 2, tail = 0)

# This returns "None"
smartEllipsis(character(0), 3)

}
