% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object-utils.R
\name{sort_df}
\alias{sort_df}
\title{Sort a data.frame by column}
\usage{
sort_df(x, cn, decreasing = FALSE, na.last = NA)
}
\arguments{
\item{x}{\code{\link{data.frame}} on which to perform the sort}

\item{cn}{\code{\link{character}} the name of the column(s) to sort.}

\item{decreasing}{\code{\link{logical}} should the sort be increasing or decreasing?}

\item{na.last}{\code{\link{logical}} see \code{\link{logical}};
for controlling the treatment of \code{NA}s.
   If \code{TRUE}, missing values in the data are put last; if \code{FALSE},
   they are put first; if \code{NA}, they are removed.}
}
\value{
\code{\link{character}}
}
\description{
Sorts the columns of a \code{\link{data.frame}}
}
\examples{
	 df<- data.frame(Name=c("one","two","three"), Value=1:3)
	 sort_df(df, cn="Value", decreasing=TRUE)
	 # sort first on column "Value", then on column "Name"
	 # note that both columns are sorted with decreasing=TRUE
	 sort_df(df, cn=c("Value","Name"), decreasing=TRUE)

}
\seealso{
\code{\link{sort}},
         \code{\link{data.frame}}
}
\author{
Thomas P. Harte (originated by John Laing)
}
\keyword{\code{\link{sort}}}
