% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prices.R
\name{stock_price}
\alias{stock_price}
\title{stock_price}
\usage{
stock_price(sec = "STK", sym, currency, exchange = "SMART", reqType = 4)
}
\arguments{
\item{sec}{security type, equals \code{STK} by default.
Other values are \code{IND} (for index), or \code{FUT} (for future)}

\item{sym}{string - IBKR style of ticker, if unknown then function returns -1}

\item{currency}{string with possibles values "USD", "CHF", "EUR".}

\item{exchange}{string - default value is "SMART". Can be also "EUREX", "CBOE",...}

\item{reqType}{default value for IBKR ticker request.}
}
\value{
a value
}
\description{
For a given ticker this function returns the last known value from IBKR. It first looks for last value stored in prices.csv file.
If one exists that is no older than 1 hour, then this value is returned. Otherwise it looks for IBKR service to retrieve a new data.
If IBKR service is not available, it will request it from end-user on console.In the end the new price is stored in prices.csv file.
}
\details{
This function is not vectorized and accepts only one ticker at a time.
}
\examples{
stock_price(sym="SPY",currency="USD")
}
