% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_api_organisations.R
\name{t_api_organisations}
\alias{t_api_organisations}
\title{Use the NHS ORD API to fetch NHS organisations}
\usage{
t_api_organisations(role = "NHS TRUST", include_inactive = FALSE)
}
\arguments{
\item{role}{character. One of c("NHS TRUST", "NHS TRUST SITE", "ICB", "SUB ICB LOCATION", "GP"). Defaults to "NHS TRUST".}

\item{include_inactive}{logical. Include inactive trusts in the
results? Defaults to FALSE.}
}
\value{
A dataframe of organisations.
}
\description{
Queries the NHS Organisation Data Service API (Organisation Reference Data - ORD)
https://digital.nhs.uk/developer/api-catalogue/organisation-data-service-ord
}
