% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_as_fy.R
\name{t_as_fy}
\alias{t_as_fy}
\title{Convert a date to the financial year it sits within}
\usage{
t_as_fy(date)
}
\arguments{
\item{date}{date. The date you are converting}
}
\value{
string. A string of the form "2020/21"
}
\description{
Convert a date to the financial year it sits within
}
