% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tau_warn.R
\name{tau_warn}
\alias{tau_warn}
\title{Log a warning message}
\usage{
tau_warn(..., sep = " ", call. = FALSE, immediate. = TRUE)
}
\arguments{
\item{...}{The message elements (characters)}

\item{sep}{Message separator (character)}

\item{call.}{Obsolete, not used.}

\item{immediate.}{Obsolete, not used.}
}
\value{
Invisible list of logger objects,
but mostly called for it's side effect.
}
\description{
This is a wrapper for logger::log_warn,
letting the caller set the separator.
}
