% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tauReadCsv.R
\name{tauReadCsv}
\alias{tauReadCsv}
\title{Tau wrapper for read.csv()}
\usage{
tauReadCsv(
  file,
  header = TRUE,
  stringsAsFactors = FALSE,
  comment.char = "#",
  ...
)
}
\arguments{
\item{file}{A connection or filename for the CSV file}

\item{header}{Passed to read.csv(), defaults to TRUE}

\item{stringsAsFactors}{Passed to read.csv(), defaults to FALSE}

\item{comment.char}{Passed to read.csv, defaults to "#"}

\item{...}{Any other arguments for read.csv()}
}
\value{
The data frame from the CSV file
}
\description{
This just calls \link{read.csv}, but it provides better defaults.
}
