% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{the}
\alias{the}
\title{Return first (and only) element of argument}
\usage{
the(x)
}
\arguments{
\item{x}{A vector or list}
}
\value{
First element of x
}
\description{
Returns first element of argument.
Stops if argument does not have exactly one element.
}
