% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/businessDates.R
\name{thisTradingDay}
\alias{thisTradingDay}
\title{The current trading date}
\usage{
thisTradingDay()
}
\value{
Date object
}
\description{
Returns the current trading date;
that is, the date for which we can trade right now.

Normally, today is the current trading date;
but on weekends, the current trading date
is typically Monday (or Tuesday, if Monday is a holiday).
}
\seealso{
\link{prevTradingDay} for the trading date before the current trading date.
}
