% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.xts.R
\name{tidy.xts}
\alias{tidy.xts}
\title{Convert xts object to tibble}
\usage{
\method{tidy}{xts}(x, pivot_longer = TRUE, ...)
}
\arguments{
\item{x}{An xts object}

\item{pivot_longer}{If TRUE, pivot columns of \code{x}
into long format}
}
\value{
Returns a tibble.
  If \code{pivot_longer} is TRUE,
  the columns are \code{index}, \code{series}, and \code{value},
  where values of \code{series} are taken from the column names of \code{x}.
  Otherwise, the tibble has the columns of \code{x},
  plus an additional column called \code{index}.
}
\description{
Convert xts object to tibble
}
