% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts-utils.R
\name{to_ISO_8601}
\alias{to_ISO_8601}
\title{Coerce character strings to the Date class}
\usage{
to_ISO_8601(strs, year.left = TRUE, american = FALSE)
}
\arguments{
\item{strs}{\code{\link{character}} string to convert to ISO 8601}

\item{year.left}{\code{\link{logical}} give \code{to_ISO_8601} a hint: does
the year occur on the left \code{TRUE} or on the right \code{FALSE}}

\item{american}{\code{\link{logical}} does the string follow the bizarre American
convention for writing dates, viz. MM-DD-YYYY}
}
\value{
\code{\link{Date}} \code{\link{vector}}
}
\description{
Coerce character strings to Date class by judiciously guessing the strings' format
}
\examples{
   require(zoo)
   require(magrittr)
   require(dplyr)
   tab<-  tutils:::make_ISO_8601_test_table()
   tab \%<>\% mutate(`Output`=as.character(NA))

   for (row in 1:nrow(tab)) {
       tab[row, "Output"]<- to_ISO_8601(
           str=tab[row, "Test"],
           year.left=tab[row, "year.left"],
           american=tab[row, "american"]
       ) \%>\% as.character
   }

   (tab \%<>\% mutate(`success`=`Expected`==`Output`))

}
\seealso{
\code{\link{Date}}, \url{https://xkcd.com/1179}
}
\author{
Thomas P. Harte
}
\keyword{\code{\link{Date}}}
