% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posts.R
\name{todaysAlerts}
\alias{todaysAlerts}
\title{All alerts generated today}
\usage{
todaysAlerts(...)
}
\arguments{
\item{...}{Passed to [\link{loadEvents}]}
}
\description{
All alerts generated today
}
\seealso{
\link{postAlert} and \link{latestAlerts}
}
