% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posts.R
\name{todaysEvents}
\alias{todaysEvents}
\title{All events generated today}
\usage{
todaysEvents(...)
}
\arguments{
\item{...}{Passed to [\link{loadEvents}]}
}
\value{
See \link{loadEvents}, the underlying function
}
\description{
All events generated today
}
\seealso{
\link{todaysEvents} for events that happened today
}
