% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{Tone}
\alias{Tone}
\alias{is.Tone}
\alias{as.Tone}
\title{Tone type}
\usage{
Tone(x)

is.Tone(x)

as.Tone(x)
}
\arguments{
\item{x}{A character vector of the values "bull" and "bear"}
}
\description{
A \code{Tone} type indicates if a market is
generally trending \emph{upward} ("bull") or \emph{downward} ("bear").
}
\seealso{
\link{Side} type, \link{Polarity} type
}
