% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string-utils.R
\name{trim}
\alias{trim}
\alias{triml,}
\alias{trimr}
\alias{triml}
\title{Trim whitespace}
\usage{
trim(str)

triml(str)

trimr(str)
}
\arguments{
\item{str}{\code{\link{character}} string}
}
\value{
\code{\link{character}}
}
\description{
Trim whitespace from ends of \code{\link{character}} string
}
\examples{
# left trim:
cat(sprintf("'\\\%s'\\n", triml(" asdf qwerty \\t")))
cat(sprintf("'\\\%s'\\n", triml("\\tasdf qwerty \\t")))
# right trim:
cat(sprintf("'\\\%s'\\n", trimr(" asdf qwerty \\t")))
cat(sprintf("'\\\%s'\\n", trimr("\\tasdf qwerty \\t ")))
# left & right trim:
cat(sprintf("'\\\%s'\\n", trim(" asdf qwerty \\t")))
cat(sprintf("'\\\%s'\\n", trim("\\tasdf qwerty \\t")))
cat(sprintf("'\\\%s'\\n", trim(" asdf qwerty \\t")))
cat(sprintf("'\\\%s'\\n", trim("\\tasdf qwerty \\t ")))

}
\seealso{
\code{\link{gsub}}
}
\author{
Thomas P Harte
}
\keyword{\code{\link{sort}}}
