% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimean.R
\name{trimean}
\alias{trimean}
\title{Trimean - robust location statistic}
\usage{
trimean(x, na.rm = FALSE)
}
\arguments{
\item{x}{Vector of numeric values}

\item{na.rm}{If TRUE, remove NA values from x}
}
\value{
Trimean value
}
\description{
The trimean is defined as (Q1 + 2*Q2 + Q3) / 4.
}
