% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_ts.R
\name{ts_ts}
\alias{ts_ts}
\title{Timestamp}
\usage{
ts_ts()
}
\value{
string in the format "year month day _ hour minute second"
}
\description{
Returns a timestamp in a format convenient for filename versioning
}
