% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tutils-package.R
\docType{package}
\name{tutils}
\alias{tutils}
\alias{tutils-package}
\title{tutils: Thomas's utility functions to make life easy (or easier)}
\description{
This package contains some functions that the author has
developed over the years and now can't do without.
Functions such as \code{\link{whos}} are essential to remain sane.
}
\references{
\url{https://github.com/tharte/tutils}
}
\author{
Thomas P. Harte \cr \url{https://github.com/tharte} \cr \cr

Thomas P. Harte \email{tharte@cantab.net} \cr
}
\keyword{package}
