% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account.R
\name{twr}
\alias{twr}
\title{twr}
\usage{
twr(dates, e_nlv, cashflows)
}
\arguments{
\item{dates}{list of dates when data are provided - there should be only one data point per date}

\item{e_nlv}{End of day net liquidation values: D1, D2, ... Dn}

\item{cashflows}{Begin of day cash flow inflows: D1, D2,... Dn}
}
\value{
a numerical vector of TWR values
}
\description{
This function Time Weighted Return computes TWR for
a given list of dates with corresponding end of day net liquidation values plus cashflows (inflows/outflows).
}
\details{
It is assumed that portfolio value begin of day (n) = portfolio value end of day (n-1)
Also it is assumed that portfolio is valued every calendar day (incl. non business days)
twr will be computed for all dates and then filtered so that it fit with input dates.
}
