% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wrapper.R
\name{unwrap}
\alias{unwrap}
\alias{unwrap.default}
\alias{unwrap.Wrapper}
\title{Extract the value from a Wrapper object}
\usage{
unwrap(x, ...)

\method{unwrap}{default}(x, ...)

\method{unwrap}{Wrapper}(x, quiet = TRUE)
}
\arguments{
\item{x}{Can be a Wrapper object or anything else}

\item{quiet}{(boolean)}
}
\value{
If x is a Wrapper object, its value is returned.
Otherwise, x is returned unchanged.
}
\description{
Extract the value from a Wrapper object
}
