% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{upsideDown}
\alias{upsideDown}
\title{Flip a matrix or data frame, top to bottom}
\usage{
upsideDown(m)
}
\arguments{
\item{m}{A matrix or data frame}
}
\value{
The input, but with rows in opposite order
}
\description{
Flip a matrix or data frame, top to bottom
}
