% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weekFactor.R
\name{weekFactor}
\alias{weekFactor}
\title{Create a week-of-the-year factor from zoo or xts object}
\usage{
weekFactor(z)
}
\arguments{
\item{z}{A zoo or xts object}
}
\value{
A factor with levels 1, ..., 53.
}
\description{
Create a week-of-the-year factor from zoo or xts object
}
