% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object-utils.R
\name{whos}
\alias{whos}
\title{List objects in an environment}
\usage{
whos(all = TRUE, envir = parent.frame(), sort.by = "Bytes",
  decreasing = FALSE, omit.classes = "function")
}
\arguments{
\item{all}{\code{\link{logical}} list all objects rather than the first 10}

\item{envir}{\code{\link{environment}} the environment in which to list objects}

\item{sort.by}{\code{\link{character}} sort on one of \code{Class},
\code{Dimensions}, \code{Bytes} or by \code{Name}}

\item{decreasing}{\code{\link{logical}} should the sort.by be increasing or decreasing?}

\item{omit.classes}{\code{\link{character}} type of objects to omit from list}
}
\value{
A \code{\link{data.frame}}, with \code{\link{row.names}} being the
  objects in the \code{\link{environment}}
}
\description{
Lists objects in a specified environment in a way that is more
useful than \code{\link{ls}}
}
\examples{
 # list objects in the current environment
 whos()

 # list objects in alphabetical order
 whos(sort="Name")

 # show all objects (specifically, include functions)
 whos(sort="Name", omit=NULL)

}
\seealso{
\code{\link{ls}}, \code{\link{sort_df}}, \code{\link{parent.frame}},
  \code{\link{environment}}
}
\author{
Thomas P. Harte
}
\keyword{\code{\link{ls}}}
