% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wrapper.R
\name{wrap}
\alias{wrap}
\title{Wrap an object}
\usage{
wrap(x, ...)
}
\arguments{
\item{value}{The value to be wrapped}

\item{w}{A Wrapper object whose value will be wrapped
and whose messages will be incorporated into the Wrapper}
}
\value{
Returns a Wrapper object
}
\description{
Wrap an object
}
\seealso{
\link{Wrapper} object class
}
