% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapObject.R
\name{wrapObject}
\alias{wrapObject}
\title{Wrap a data frame to create an object instance}
\usage{
wrapObject(df, className, required = NULL, optional = NULL, cleanse = FALSE)
}
\arguments{
\item{df}{Data frame to be wrapped (data.frame)}

\item{className}{Class of new instance (character)}

\item{required}{Names of required columns, NULL if none (character, optional)}

\item{optional}{Name of optional columns, NULL if none (character, optional)}

\item{cleanse}{If TRUE, remove all but the required and optional columns (logical)}
}
\value{
The input data frame, converted to a tibble, possibly
cleansed of extraneous columns (\code{cleanse = TRUE}),
and with a class attribute given by \code{className}.
}
\description{
Wrap a data frame to create an object instance
}
\examples{
df <- data.frame(Pi = pi, E = exp(1), Foo = "fum")
wrapObject(df, "Numbers")
wrapObject(df, "Numbers", required = c("Pi"), cleanse = TRUE)
wrapObject(df, "Numbers", required = c("Pi"), optional = c("E"))
wrapObject(df, "Numbers", required = c("Pi"), optional = c("E"), cleanse = TRUE)

}
