% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtsConversion.R
\name{xtsToDataFrame}
\alias{xtsToDataFrame}
\title{Convert xts matrix to data frame}
\usage{
xtsToDataFrame(x)
}
\arguments{
\item{x}{An xts matrix}
}
\value{
A data frame with the input columns
   plus a new 'date' column
}
\description{
DEPRECATED. Use tidy() instead.
}
