% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtsConversion.R
\name{xtsToRegressionData}
\alias{xtsToRegressionData}
\title{Convert xts to regression-ready data frame}
\usage{
xtsToRegressionData(x)
}
\arguments{
\item{x}{An xts matrix}
}
\value{
A data frame with columns
   \code{date}, \code{symbol}, and \code{value}
}
\description{
DEPRECATED. Use tidy() instead.
}
\details{
The three output columns are:
 \itemize{
     \item \code{date} - Taken from xts index
     \item \code{symbol} - Taken from xts column names
     \item \code{value} - Taken from the matrix cell at the intersection
        of \code{date} and \code{symbol}
 }
}
