% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yes_no.R
\name{yes_no}
\alias{yes_no}
\title{Yes or no?}
\usage{
yes_no(b)
}
\arguments{
\item{b}{A boolean value}
}
\value{
Returns "yes" if b is a single, defined boolean value that is TRUE;
"no" otherwise
}
\description{
Yes or no?
}
