% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateFunctions.R
\name{yyyymmddToDate}
\alias{yyyymmddToDate}
\title{Convert a yyyymmdd value to a Date}
\usage{
yyyymmddToDate(yyyymmdd)
}
\arguments{
\item{yyyymmdd}{A string or integer of year, month, date}
}
\value{
A Date object
}
\description{
Convert a yyyymmdd value to a Date
}
