% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/businessDates.R
\name{businessCalendar}
\alias{businessCalendar}
\title{Generate vector of business dates}
\usage{
businessCalendar(from = START_OF_TIME, to = Sys.Date())
}
\arguments{
\item{from}{Date or numeric;
numeric is number of days before \code{to} date}

\item{to}{Date}
}
\value{
Vector of Date objects
}
\description{
Generate vector of business dates
}
