% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/businessDates.R
\name{businessCalendar}
\alias{businessCalendar}
\title{Generate vector of business dates}
\usage{
businessCalendar(from = START_OF_TIME, to = NULL)
}
\arguments{
\item{from}{A Date object or a character string in standard date format;
or numeric, meaning the number of days before \code{to} date.
Defaults to \code{START_OF_TIME}, which is the first date
for which we have historical data.}

\item{to}{A Date object or a character string in standard date format.
Defaults to current date (\code{Sys.Date()}).}
}
\value{
Vector of Date objects
}
\description{
Given a \code{from} date and a \code{to} date, return
a vector of business days from start to end.
}
