% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{do.cbind}
\alias{do.cbind}
\title{Bind list elements into columns - OBSOLETE}
\usage{
do.cbind(lst)
}
\arguments{
\item{lst}{A list}
}
\value{
A matrix whose columns are taken from the list elements
}
\description{
OBSOLETE. Replaced by dplyr::bind_cols.
}
\details{
This is handy with the pipe operator:
 aList |> lapply(aFunction) |> do.cbind()
}
\seealso{
\link{do.rbind}
}
