% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/financialFmt.R
\name{financialFmt}
\alias{financialFmt}
\alias{dollarFmt}
\title{Format a number in "financial" format}
\usage{
financialFmt(x, prefix = "", accuracy = 1)

dollarFmt(x, ...)
}
\arguments{
\item{x}{Value to be formatted (numeric)}

\item{prefix}{Prefix to prepend, typically a currency symbol (character, optional)}

\item{accuracy}{Accuracy of formatted result;
e.g., 1 for rounding to nearest dollar, 0.01 for rounding to nearest penny
(numeric, optional)}

\item{...}{Passed to \code{financialFmr}}
}
\value{
A string
}
\description{
Format a number in "financial" format
}
