% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts-utils.R
\name{get_return}
\alias{get_return}
\title{Compute (simple net) return of values in vector}
\usage{
get_return(x, nan.replace = FALSE)
}
\arguments{
\item{x}{\code{\link{numeric}} \code{\link{vector}}}

\item{nan.replace}{\code{\link{logical}} if TRUE, replace \code{\link{NaN}} (not-a-number) symbols with \code{NA}}
}
\value{
\code{\link{numeric}} \code{\link{vector}}
}
\description{
Compute (simple net) return of values in vector
}
\examples{
  get_return(c(NA))
  get_return(c(NA, NA))
  get_return(c(NaN, NA))
  get_return(c(NaN, 1, 1.1), nan.replace=TRUE)
  get_return(c(1, +Inf, 1.1), nan.replace=TRUE)

  dates<- seq(as.Date("2011-01-01"), by=1, len=6)
 z<- zoo(c(NA, 1, 2, 3, 0, 0), order.by=dates)

  zoo::rollapply(z, 2, get_return, align="right", nan.replace=TRUE)
  # 2011-01-02 2011-01-03 2011-01-04 2011-01-05 2011-01-06
  #	  NA        1.0        0.5       -1.0         NA

  zoo::rollapply(z, 2, get.return, align="right", nan.replace=FALSE)
  # 2011-01-02 2011-01-03 2011-01-04 2011-01-05 2011-01-06
  #	  NA        1.0        0.5       -1.0         NaN

}
\seealso{
\code{\link{numeric}}, \code{\link{NaN}}, \code{\link{vector}}
}
\author{
Thomas P. Harte
}
\keyword{\code{\link{NaN}},}
\keyword{\code{\link{numeric}},}
\keyword{\code{\link{vector}}}
