% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/businessDates.R
\name{isBusinessDay}
\alias{isBusinessDay}
\title{Predicate for business days}
\usage{
isBusinessDay(dates)
}
\arguments{
\item{dates}{Vector of Date objects}
}
\value{
Logical vector
}
\description{
Predicate for business days
}
