% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datePartFactors.R, R/monthFunctions.R
\name{monthFactor}
\alias{monthFactor}
\title{Create month-of-year factor from zoo or xts object}
\usage{
monthFactor(z)

monthFactor(z)
}
\arguments{
\item{z}{The zoo or xts object}
}
\value{
A factor

Factor with levels Jan, Feb, Mar, ...
}
\description{
Given a zoo or xts object, return a month factor
 based on the index. Factor levels are Jan, Feb, Mar, ...
}
\details{
TODO: Really, there should be monthFactor.zoo,
   monthFactor.xts, monthFactor.Date, monthFactor.character, monthFactor.integer, etc.
}
\seealso{
months
}
