% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/businessDates.R
\name{nearestBusinessDay}
\alias{nearestBusinessDay}
\title{Find the business dates nearest to given dates}
\usage{
nearestBusinessDay(dates = Sys.Date())
}
\arguments{
\item{dates}{Vector of Date objects}
}
\value{
Vector of Date objects
}
\description{
Given a vector of dates, find the most recent business dates
on or before the given dates.
}
\examples{
# If today is a business day, return today.
# Otherwise, return the most recent business date.
nearestBusinessDay(Sys.Date())
}
