% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posts.R
\name{postStatus}
\alias{postStatus}
\title{Post a status message}
\usage{
postStatus(
  origin,
  status,
  message = NULL,
  url = NA_character_,
  link_text = NA_character_
)
}
\arguments{
\item{origin}{Name of program that is posting status (character)}

\item{status}{Typically "OK" or "Error" or "Failure" (character)}

\item{message}{Useful, descriptive text message for user
(optional, character)}

\item{url}{URL of related app page, if any
(optional, character)}

\item{link_text}{Link text for anchor of related app page, if any
(optional, character)}
}
\value{
Nothing
}
\description{
Post a status message
}
\seealso{
See \link{readStatus} for reading status messages.
}
