% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smartEllipsis.R
\name{smartEllipsis}
\alias{smartEllipsis}
\title{Employ an ellipsis, if necessary}
\usage{
smartEllipsis(strings, max = 2, tail = 1, empty = "None")
}
\arguments{
\item{strings}{Vector of strings to be joined (character)}

\item{max}{Number of strings before ellipsis (integer)}

\item{tail}{Number of strings after the ellipsis (integer, optional)}

\item{empty}{Returned if \code{strings} is empty (character)}
}
\value{
All the \code{strings} pasted together, separated by commas,
  up to a maximum of \code{max} strings;
  if \code{strings} has more than \code{max} elements,
  the extra elements are replaced by an ellipsis.
  If \code{strings} is empty, then \code{empty} is returned.
}
\description{
Employ an ellipsis, if necessary
}
\examples{
smartEllipsis(c("tom", "dick", "harry", "peter"), 4)
smartEllipsis(c("tom", "dick", "harry", "peter"), 3)
smartEllipsis(c("tom", "dick", "harry", "peter"), 3, tail = 0)
smartEllipsis(c("tom", "dick", "harry", "peter"), 2)
smartEllipsis(c("tom", "dick", "harry", "peter"), 2, tail = 0)
smartEllipsis(character(0), 3)

}
