% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tau_warn.R
\name{tau_warn}
\alias{tau_warn}
\title{Tau-style warning message}
\usage{
tau_warn(..., sep = " ", call. = FALSE, immediate. = TRUE)
}
\arguments{
\item{...}{The message (characters)}

\item{sep}{Message separator (character)}

\item{call.}{Passed to base::warning (logical)}

\item{immediate.}{Passed to base::warning (logical)}
}
\value{
Returns the warning message as a character string, invisibly,
but mostly called for its side effect
}
\description{
This is like base::warning(), but friendlier.
}
