% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wrapper.R
\name{Wrapper}
\alias{Wrapper}
\alias{wrap}
\alias{unwrap}
\alias{unwrap.default}
\alias{unwrap.Wrapper}
\title{Wrap a value and optional messages}
\usage{
Wrapper(value, messages, ...)

wrap(value, messages = list(), ...)

unwrap(x, ...)

\method{unwrap}{default}(x, ...)

\method{unwrap}{Wrapper}(x, element = NULL, quiet = TRUE)
}
\arguments{
\item{messages}{List of associated messages (list, optional)}

\item{x}{A value to be wrapped}

\item{element}{(character)}

\item{quiet}{If TRUE, print the enclosed messages, if any, before
extracting the enclosed value (logical)}
}
\value{
Return a list with class \code{Wrapper}
}
\description{
The \code{wrap} functions will, uh, wrap a value inside a container object,
 possibly with some messages.
}
\details{
The \code{unwrap} function extracts the value from the container,
 possibly printing the messages, if any.
}
\examples{
w <- wrap(pi, messages = "Help me!")

# Extract the wrapped value
unwrap(w)

# Extract the wrapped value and print the wrapped messages
unwrap(w, quiet = FALSE)


}
