% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wrapper.R
\name{Wrapper}
\alias{Wrapper}
\alias{is.Wrapper}
\alias{wrap}
\alias{unwrap}
\alias{unwrap.default}
\alias{unwrap.Wrapper}
\title{Wrap a value and optional messages}
\usage{
Wrapper(value, errors = list, warnings = list(), info = list(), ...)

is.Wrapper(x)

wrap(value, ...)

unwrap(x, ...)

\method{unwrap}{default}(x, ...)

\method{unwrap}{Wrapper}(x, element = NULL, quiet = TRUE)
}
\arguments{
\item{errors}{List of error messages (list, optional)}

\item{warnings}{List of warning messages (list, optional)}

\item{info}{List of informational messages (list, optional)}

\item{x}{A value to be wrapped}

\item{element}{(character)}

\item{quiet}{If TRUE, print the enclosed messages, if any, before
extracting the enclosed value (logical)}
}
\value{
Return a list with class \code{Wrapper}
}
\description{
The \code{wrap} function wraps a value inside a special container object,
 possibly wrapping some messages.
}
\details{
The \code{unwrap} function extracts the value from the container.
 Unless \code{quiet = TRUE}, it will print the wrapped, associated messages, if any,
 before returning that extracted value.
}
\examples{
w <- wrap(pi, error = "Help me!", warnings = "Just chillin'")

# Extract the wrapped value
unwrap(w)

# Extract the wrapped value and print the wrapped messages
unwrap(w, quiet = FALSE)

}
